refactor(observability): 分离体验报告编解码
