+++ src/observability/experience.ts
@@ -1,4 +1,3 @@
-import { isTraceSourceKind as isObservationSourceKind } from '../shared/trace-source-kind.js';
 import type {
   ExperienceAssistiveInference,
   ExperienceAssistiveInferenceCautionCode,
@@ -63,10 +62,8 @@ import type {
   ExperienceStoryContext,
   ExperienceTimelineBranch,
   ExperienceTimelineEvent,
-  ExperienceTraceRecordRange,
   ExperienceTraceTimeline,
   ExperienceTimelineTree,
-  ExperienceTurnSummary,
   ObservationExperienceReport,
 } from './contracts/experience.js';
 import type { ObservationInboxItem } from './contracts/inbox.js';
@@ -79,7 +76,6 @@ import {
 import { observationMetricAnnotationVerdict, observationReviewStateKey } from './review-state.js';
 import {
   correlateTraceToolEvents,
-  normalizeTraceTimestamp,
   type TraceSession,
 } from './trace-ir.js';
 import {
@@ -127,7 +123,6 @@ import {
   enrichRouterDownstreamIndicators,
   evidenceRefFromTimeline,
   invocationTimestampObserved,
-  isAssistantProgressUpdateEvent,
   primarySourceTraceForSession,
   priorityForReviewerFindings,
   priorityForScore,
@@ -174,6 +169,17 @@ export {
 export type { TextMatchRange } from './feedback-matchers.js';
 export { projectTraceSessionTimeline } from './experience/timeline.js';
 export { OBSERVATION_EXPERIENCE_SCHEMA_VERSION };
+export {
+  compactObservationExperienceReport,
+  normalizeObservationExperienceReport,
+} from './experience/report-codec.js';
+export type {
+  PersistedExperienceInvocation,
+  PersistedExperienceReviewerReport,
+  PersistedExperienceSession,
+  PersistedExperienceSessionStory,
+  PersistedObservationExperienceReport,
+} from './experience/report-codec.js';
 
 export type {
   ExperienceAssistiveInference,
@@ -243,54 +249,6 @@ export type {
   ExperienceTimelineTree,
   ObservationExperienceReport,
   ObservationReviewState,
-};
-const LEGACY_OBSERVATION_EXPERIENCE_SCHEMA_VERSION = 2;
-
-export type PersistedExperienceInvocation = Omit<
-  ExperienceInvocation,
-  'timeline' | 'timelineRef' | 'timelineEventIds'
-> & {
-  timelineRef: string;
-  timelineEventIds: string[];
-};
-
-export type PersistedExperienceSessionStory = Omit<
-  ExperienceSessionStory,
-  'contextRef' | 'goalSlices' | 'subagentDispatches' | 'episodes'
-> & {
-  contextRef: string;
-};
-
-export type PersistedExperienceReviewerReport = Omit<
-  ExperienceReviewerReport,
-  'sessionStory' | 'sessionStoryRef'
-> & {
-  sessionStoryRef: 'session';
-};
-
-export type PersistedExperienceSession = Omit<
-  ExperienceSessionSummary,
-  | 'timelineRef'
-  | 'timelinePreviewEventIds'
-  | 'attributedEventIds'
-  | 'timelinePreview'
-  | 'fullSessionTimeline'
-  | 'timelineTree'
-  | 'sessionStory'
-  | 'reviewerReport'
-> & {
-  timelineRef: string;
-  timelinePreviewEventIds: string[];
-  sessionStory?: PersistedExperienceSessionStory;
-  reviewerReport?: PersistedExperienceReviewerReport;
-};
-
-export type PersistedObservationExperienceReport = Omit<
-  ObservationExperienceReport,
-  'invocations' | 'sessions'
-> & {
-  invocations: PersistedExperienceInvocation[];
-  sessions: PersistedExperienceSession[];
 };
 
 export function aggregateExperienceChecklistItemStatus(statuses: ExperienceChecklistItemStatus[]): ExperienceChecklistItemStatus {
@@ -445,2414 +403,11 @@ export function buildObservationExperienceReport(input: BuildExperienceInput): O
   };
 }
 
-export function normalizeObservationExperienceReport(value: unknown): ObservationExperienceReport | null {
-  if (!value || typeof value !== 'object') return null;
-  const report = value as Record<string, unknown>;
-  const kind = report.kind === 'observe-experience' ? report.kind : null;
-  if (!kind) return null;
-  if (
-    report.schemaVersion !== OBSERVATION_EXPERIENCE_SCHEMA_VERSION
-    && report.schemaVersion !== LEGACY_OBSERVATION_EXPERIENCE_SCHEMA_VERSION
-  ) return null;
-  const isLegacyReport = report.schemaVersion === LEGACY_OBSERVATION_EXPERIENCE_SCHEMA_VERSION;
-  if (report.scope !== 'evidence-only') return null;
-  if (
-    !isTimestamp(report.generatedAt)
-    || !report.meta
-    || typeof report.meta !== 'object'
-    || !isExperienceMeta(report.meta)
-  ) return null;
-  if (!Array.isArray(report.goalSlices) || !Array.isArray(report.invocations) || !Array.isArray(report.sessions) || !Array.isArray(report.skills)) {
-    return null;
-  }
-  const invocations = normalizeExperienceInvocationShells(report.invocations, !isLegacyReport);
-  const sessions = normalizeExperienceSessionShells(report.sessions, !isLegacyReport);
-  if (!invocations || !sessions) return null;
-  if (
-    !isLegacyReport
-    && (!Array.isArray(report.traceTimelines) || !Array.isArray(report.storyContexts))
-  ) return null;
-  const traceTimelines = Array.isArray(report.traceTimelines)
-    ? normalizeTraceTimelines(report.traceTimelines)
-    : traceTimelinesFromSessions(sessions, invocations);
-  const storyContexts = Array.isArray(report.storyContexts)
-    ? normalizeStoryContexts(report.storyContexts)
-    : storyContextsFromSessions(sessions, invocations);
-  if (!traceTimelines || !storyContexts) return null;
-  const normalized: ObservationExperienceReport = {
-    kind: 'observe-experience',
-    schemaVersion: OBSERVATION_EXPERIENCE_SCHEMA_VERSION,
-    scope: 'evidence-only',
-    generatedAt: report.generatedAt,
-    meta: report.meta as ObservationExperienceReport['meta'],
-    goalSlices: report.goalSlices as ObservationExperienceReport['goalSlices'],
-    traceTimelines,
-    storyContexts,
-    invocations,
-    sessions,
-    skills: report.skills as ObservationExperienceReport['skills'],
-  };
-  const hydrated = hydrateExperienceTimelines(normalized);
-  try {
-    if (!validateExperienceReferences(hydrated, !isLegacyReport)) return null;
-  } catch {
-    return null;
-  }
-  return hydrated;
-}
-
-function normalizeExperienceInvocationShells(
-  values: unknown[],
-  strict: boolean,
-): ExperienceInvocation[] | null {
-  const records = values.filter(isObjectRecord);
-  if (records.length !== values.length) return null;
-  if (records.some((value) =>
-    typeof value.id !== 'string'
-    || typeof value.skillName !== 'string'
-    || typeof value.sessionId !== 'string'
-    || typeof value.sessionGroupKey !== 'string'
-    || (strict ? typeof value.traceId !== 'string' : !isOptionalString(value.traceId))
-    || typeof value.sourceTrace !== 'string'
-    || !isObservationSourceKind(value.sourceKind)
-    || !isOptionalString(value.entrypoint)
-    || !isOptionalTraceSourceMetadata(value.sourceMetadata)
-    || !isOptionalString(value.cwd)
-    || !isNonNegativeInteger(value.segmentIndex)
-    || typeof value.goalSliceId !== 'string'
-    || !isTimestampRange(value.startTimestamp, value.endTimestamp)
-    || (value.timestampObserved !== undefined && typeof value.timestampObserved !== 'boolean')
-    || (strict && typeof value.timestampObserved !== 'boolean')
-    || !isExperienceAttribution(value.attribution)
-    || !isExperienceInvocationMetrics(value.metrics, strict)
-    || !isCountRecord(value.toolCounts)
-    || !isExperienceIndicators(value.indicators, strict)
-    || !isExperienceEvidenceChain(value.evidenceChain)
-    || !isExperienceRuleFindingArray(value.ruleFindings)
-    || !isExperienceAssistiveInference(value.assistiveInference)
-    || !isExperienceProblemPatternArray(value.problemPatterns)
-    || !isStringArray(value.relatedObservationIds)
-    || !isExperienceEvidenceRefArray(value.evidenceRefs)
-    || (
-      strict
-        ? typeof value.timelineRef !== 'string'
-          || !isStringArray(value.timelineEventIds)
-          || (value.timeline !== undefined && !isTimelineEventArray(value.timeline))
-        : !isOptionalString(value.timelineRef)
-          || (value.timelineEventIds !== undefined && !isStringArray(value.timelineEventIds))
-          || !isTimelineEventArray(value.timeline)
-    )
-  )) return null;
-  return records.map((value) => ({
-    ...value,
-    metrics: {
-      ...(value.metrics as Record<string, unknown>),
-      tokenUsageObserved: strict
-        ? (value.metrics as Record<string, unknown>).tokenUsageObserved
-        : false,
-    },
-    timeline: isTimelineEventArray(value.timeline)
-      ? value.timeline as ExperienceTimelineEvent[]
-      : [],
-  } as ExperienceInvocation));
-}
-
-function normalizeExperienceSessionShells(
-  values: unknown[],
-  strict: boolean,
-): ExperienceSessionSummary[] | null {
-  const records = values.filter(isObjectRecord);
-  if (records.length !== values.length) return null;
-  if (records.some((value) =>
-    typeof value.id !== 'string'
-    || typeof value.skillName !== 'string'
-    || !isOptionalString(value.threadId)
-    || !isOptionalString(value.sourceThreadId)
-    || typeof value.sessionId !== 'string'
-    || typeof value.sourceTrace !== 'string'
-    || !isObservationSourceKind(value.sourceKind)
-    || !isOptionalString(value.entrypoint)
-    || !isOptionalTraceSourceMetadata(value.sourceMetadata)
-    || !isOptionalString(value.cwd)
-    || !isConsistentSourceSessionTime(value)
-    || !isTimestampRange(value.startTimestamp, value.endTimestamp)
-    || !isStringArray(value.invocationIds)
-    || !isOptionalTimestampCoverage(
-      value.timestampedInvocationCount,
-      value.timestampCoverage,
-      Array.isArray(value.invocationIds) ? value.invocationIds.length : -1,
-    )
-    || (
-      strict
-      && (
-        !isNonNegativeInteger(value.timestampedInvocationCount)
-        || !isRate(value.timestampCoverage)
-      )
-    )
-    || !isStringArray(value.goalSliceIds)
-    || !isExperienceReviewPriority(value.reviewPriority)
-    || typeof value.reviewPriorityScore !== 'number'
-    || !Number.isFinite(value.reviewPriorityScore)
-    || value.reviewPriorityScore < 0
-    || !isEnumArray(value.reviewBasisCodes, [
-      'has_high_observation',
-      'has_medium_observation',
-      'user_correction',
-      'user_interruption',
-      'session_interrupted',
-      'negative_feedback',
-      'hard_rule_text_hit',
-      'tool_failure',
-      'hedging_signal',
-      'explicit_marker',
-    ])
-    || !isExperienceIndicators(value.indicators, strict)
-    || !isExperienceEvidenceChain(value.evidenceChain)
-    || !isExperienceRuleFindingArray(value.ruleFindings)
-    || !isExperienceAssistiveInference(value.assistiveInference)
-    || !isExperienceProblemPatternArray(value.problemPatterns)
-    || !isStringArray(value.relatedObservationIds)
-    || (value.turns !== undefined && !isExperienceTurnSummaryArray(value.turns))
-    || !(strict
-      ? isExperienceTimelineScope(value.timelineScope)
-      : isLegacyExperienceTimelineScope(value.timelineScope))
-    || !isStringArray(value.attributionSources)
-    || !isStringArray(value.pluginNames)
-    || !isStringArray(value.rawSkillRefs)
-    || !isStringArray(value.commandNames)
-    || (
-      strict
-        ? typeof value.timelineRef !== 'string'
-          || !isStringArray(value.timelinePreviewEventIds)
-          || (value.timelinePreview !== undefined && !isTimelineEventArray(value.timelinePreview))
-          || (value.fullSessionTimeline !== undefined && !isTimelineEventArray(value.fullSessionTimeline))
-        : !isOptionalString(value.timelineRef)
-          || (
-            value.timelinePreviewEventIds !== undefined
-            && !isStringArray(value.timelinePreviewEventIds)
-          )
-          || !isTimelineEventArray(value.timelinePreview)
-          || !isTimelineEventArray(value.fullSessionTimeline)
-    )
-    || (value.timelineTree !== undefined && !isTimelineTree(value.timelineTree))
-    || (
-      value.sessionStory !== undefined
-      && !isExperienceSessionStory(value.sessionStory, strict)
-    )
-    || (
-      value.reviewerReport !== undefined
-      && !isExperienceReviewerReport(value.reviewerReport, strict)
-    )
-  )) return null;
-  return records.map((value) => {
-    const sourceThreadId = typeof value.sourceThreadId === 'string'
-      ? value.sourceThreadId
-      : typeof value.threadId === 'string'
-        ? value.threadId
-        : value.sessionId as string;
-    const threadId = typeof value.threadId === 'string'
-      ? value.threadId
-      : hashParts('thread', sourceThreadId);
-    const sessionStory = isObjectRecord(value.sessionStory)
-      ? {
-          ...value.sessionStory,
-          goalSlices: Array.isArray(value.sessionStory.goalSlices) ? value.sessionStory.goalSlices : [],
-          subagentDispatches: Array.isArray(value.sessionStory.subagentDispatches)
-            ? value.sessionStory.subagentDispatches
-            : [],
-          episodes: Array.isArray(value.sessionStory.episodes) ? value.sessionStory.episodes : [],
-        }
-      : undefined;
-    return {
-      ...value,
-      threadId,
-      sourceThreadId,
-      invocationIds: Array.isArray(value.invocationIds) ? value.invocationIds : [],
-      turns: isExperienceTurnSummaryArray(value.turns) ? value.turns : [],
-      timelinePreview: Array.isArray(value.timelinePreview) ? value.timelinePreview : [],
-      fullSessionTimeline: Array.isArray(value.fullSessionTimeline) ? value.fullSessionTimeline : [],
-      sessionStory,
-    } as ExperienceSessionSummary;
-  });
-}
-
-function normalizeTraceTimelines(values: unknown[]): ExperienceTraceTimeline[] | null {
-  if (values.some((value) => {
-    if (
-      !isObjectRecord(value)
-      || typeof value.id !== 'string'
-      || typeof value.sessionGroupKey !== 'string'
-      || typeof value.sessionId !== 'string'
-      || !isNonNegativeInteger(value.eventCount)
-    ) return true;
-    return !isTimelineTree(value.tree);
-  })) return null;
-  return values as ExperienceTraceTimeline[];
-}
-
-function normalizeStoryContexts(values: unknown[]): ExperienceStoryContext[] | null {
-  if (values.some((value) => !isExperienceStoryContext(value))) return null;
-  return values as ExperienceStoryContext[];
-}
-
-function isNonNegativeInteger(value: unknown): value is number {
-  return typeof value === 'number' && Number.isSafeInteger(value) && value >= 0;
-}
-
-function isRate(value: unknown): value is number {
-  return typeof value === 'number' && Number.isFinite(value) && value >= 0 && value <= 1;
-}
-
-function isOptionalTimestampCoverage(
-  count: unknown,
-  coverage: unknown,
-  total: number,
-): boolean {
-  if (count === undefined && coverage === undefined) return true;
-  if (
-    total < 0
-    || !isNonNegativeInteger(count)
-    || count > total
-    || !isRate(coverage)
-  ) return false;
-  return coverage === (total > 0 ? count / total : 0);
-}
-
-function isStringArray(value: unknown): value is string[] {
-  return Array.isArray(value) && value.every((item) => typeof item === 'string');
-}
-
-function isOptionalString(value: unknown): boolean {
-  return value === undefined || typeof value === 'string';
-}
-
-function isTimestamp(value: unknown): value is string {
-  return typeof value === 'string' && normalizeTraceTimestamp(value) !== undefined;
-}
-
-function isOptionalTimestamp(value: unknown): boolean {
-  return value === undefined || isTimestamp(value);
-}
-
-function isTimestampRange(start: unknown, end: unknown): boolean {
-  if (!isTimestamp(start) || !isTimestamp(end)) return false;
-  return Date.parse(start) <= Date.parse(end);
-}
-
 function sessionTimestampedInvocationCount(session: ExperienceSessionSummary): number {
   return session.timestampedInvocationCount
     ?? (session.startTimestamp === UNOBSERVED_TRACE_TIMESTAMP
       ? 0
       : session.invocationIds.length);
-}
-
-function isConsistentSourceSessionTime(value: Record<string, unknown>): boolean {
-  const start = value.sourceSessionStartTimestamp;
-  const end = value.sourceSessionEndTimestamp;
-  const duration = value.sourceSessionDurationMs;
-  if (!isOptionalTimestamp(start) || !isOptionalTimestamp(end)) return false;
-  if (duration !== undefined && !isNonNegativeInteger(duration)) return false;
-  if (start === undefined || end === undefined) return duration === undefined;
-  return typeof start === 'string'
-    && typeof end === 'string'
-    && isTimestampRange(start, end)
-    && duration === durationMsBetween(start, end);
-}
-
-function isOptionalNonNegativeInteger(value: unknown): boolean {
-  return value === undefined || isNonNegativeInteger(value);
-}
-
-function isEnumValue(value: unknown, values: readonly string[]): boolean {
-  return typeof value === 'string' && values.includes(value);
-}
-
-function isEnumArray(value: unknown, values: readonly string[]): boolean {
-  return Array.isArray(value) && value.every((item) => isEnumValue(item, values));
-}
-
-function isExperienceMeta(value: unknown): boolean {
-  if (
-    !isObjectRecord(value)
-    || !isNonNegativeInteger(value.sessionCount)
-    || !isNonNegativeInteger(value.skillCount)
-    || !isNonNegativeInteger(value.invocationCount)
-    || !isNonNegativeInteger(value.goalSliceCount)
-  ) return false;
-  return isEnumArray(value.noteCodes, [
-    'no_llm_judge',
-    'no_auto_verdict',
-    'default_goal_slice_is_allowed',
-    'deterministic_assistive_inference',
-  ]);
-}
-
-function isOptionalTraceSourceMetadata(value: unknown): boolean {
-  if (value === undefined) return true;
-  return isObjectRecord(value)
-    && isOptionalString(value.channel)
-    && isOptionalString(value.sender)
-    && isOptionalString(value.senderId)
-    && isOptionalString(value.provider)
-    && isOptionalString(value.model)
-    && isOptionalString(value.modelApi)
-    && (value.businessActions === undefined || isStringArray(value.businessActions));
-}
-
-function isExperienceReviewPriority(value: unknown): boolean {
-  return value === 'review_first' || value === 'sample_review' || value === 'routine_sample';
-}
-
-function isExperienceEvidenceKind(value: unknown): boolean {
-  return isEnumValue(value, [
-    'user_message',
-    'synthetic_user_event',
-    'assistant_message',
-    'model_activity',
-    'agent_activity',
-    'tool_use',
-    'tool_result',
-    'skill_context',
-    'runtime_context',
-    'lifecycle',
-    'observation',
-  ]);
-}
-
-function isExperienceEvidenceRef(value: unknown): value is ExperienceEvidenceRef {
-  if (
-    !isObjectRecord(value)
-    || typeof value.id !== 'string'
-    || !isExperienceEvidenceKind(value.kind)
-    || typeof value.sourceTrace !== 'string'
-    || typeof value.sessionId !== 'string'
-    || !isOptionalNonNegativeInteger(value.messageIndex)
-    || !isOptionalNonNegativeInteger(value.logicalMessageIndex)
-    || !isOptionalNonNegativeInteger(value.sourceLineIndex)
-  ) return false;
-  if (
-    value.traceRole !== undefined
-    && !isEnumValue(value.traceRole, ['standalone', 'main', 'subagent'])
-  ) return false;
-  if (value.modelActivityKind !== undefined && value.modelActivityKind !== 'reasoning') return false;
-  if (
-    value.contentVisibility !== undefined
-    && !isEnumValue(value.contentVisibility, ['plaintext', 'opaque'])
-  ) return false;
-  if (
-    value.contentSource !== undefined
-    && !isEnumValue(value.contentSource, ['summary', 'content', 'text'])
-  ) return false;
-  if (
-    value.runtimeKind !== undefined
-    && !isEnumValue(value.runtimeKind, ['session_context', 'execution_context', 'settings', 'goal', 'context_compaction', 'usage'])
-  ) return false;
-  if (
-    value.role !== undefined
-    && !isEnumValue(value.role, ['user', 'assistant', 'tool', 'other'])
-  ) return false;
-  return [
-    value.traceLabel,
-    value.traceId,
-    value.turnId,
-    value.messageUuid,
-    value.sourceType,
-    value.callInstanceId,
-    value.toolUseId,
-    value.label,
-    value.snippet,
-  ].every(isOptionalString)
-    && isOptionalTimestamp(value.timestamp);
-}
-
-function isExperienceEvidenceRefArray(value: unknown): value is ExperienceEvidenceRef[] {
-  return Array.isArray(value) && value.every(isExperienceEvidenceRef);
-}
-
-function isExperienceGoalSlice(value: unknown): value is ExperienceGoalSlice {
-  return isObjectRecord(value)
-    && typeof value.id === 'string'
-    && typeof value.skillName === 'string'
-    && typeof value.sessionId === 'string'
-    && isOptionalString(value.traceId)
-    && typeof value.sourceTrace === 'string'
-    && isOptionalString(value.cwd)
-    && isTimestampRange(value.startTimestamp, value.endTimestamp)
-    && (value.timestampObserved === undefined || typeof value.timestampObserved === 'boolean')
-    && isEnumValue(value.sliceReasonCode, [
-      'skill_segment_boundary',
-      'explicit_user_goal_shift',
-      'default_session_slice',
-    ])
-    && isEnumValue(value.sliceConfidence, ['low', 'medium', 'high'])
-    && isOptionalString(value.inferredUserGoal)
-    && isExperienceEvidenceRefArray(value.userMessageRefs);
-}
-
-function isExperienceEvidenceChain(value: unknown): boolean {
-  if (!isObjectRecord(value)) return false;
-  const countKeys = [
-    'userMessageCount',
-    'runtimeContextCount',
-    'skillContextCount',
-    'assistantMessageCount',
-    'toolUseCount',
-    'toolResultCount',
-    'toolFailureResultCount',
-    'observationCount',
-  ];
-  if (!countKeys.every((key) => isNonNegativeInteger(value[key]))) return false;
-  const optionalRefs = [
-    value.firstUserMessage,
-    value.firstRuntimeContext,
-    value.firstSkillContext,
-    value.firstToolUse,
-    value.firstToolFailure,
-    value.lastAssistantMessage,
-  ];
-  return optionalRefs.every((ref) => ref === undefined || isExperienceEvidenceRef(ref));
-}
-
-function isExperienceRuleFinding(value: unknown): boolean {
-  return isObjectRecord(value)
-    && isEnumValue(value.code, [
-      'high_observation_seen',
-      'medium_observation_seen',
-      'user_correction_seen',
-      'user_interruption_seen',
-      'session_interrupted_seen',
-      'negative_feedback_seen',
-      'positive_feedback_seen',
-      'user_goal_shift_seen',
-      'hard_rule_seen',
-      'tool_failure_seen',
-      'hedging_seen',
-      'explicit_marker_seen',
-      'runtime_context_excluded',
-      'skill_context_excluded',
-      'no_priority_signal',
-    ])
-    && isEnumValue(value.level, ['attention', 'sample', 'normal'])
-    && isNonNegativeInteger(value.count)
-    && isExperienceEvidenceRefArray(value.evidenceRefs);
-}
-
-function isExperienceRuleFindingArray(value: unknown): boolean {
-  return Array.isArray(value) && value.every(isExperienceRuleFinding);
-}
-
-function isExperienceAssistiveInference(value: unknown): boolean {
-  return isObjectRecord(value)
-    && value.mode === 'deterministic_rules_only'
-    && isEnumValue(value.code, [
-      'review_recommended',
-      'sample_recommended',
-      'positive_signal_observed',
-      'user_switched_topic_neutral',
-      'no_obvious_issue_from_rules',
-      'insufficient_human_context',
-    ])
-    && isEnumValue(value.confidence, ['low', 'medium', 'high'])
-    && isEnumArray(value.basisRuleCodes, [
-      'high_observation_seen',
-      'medium_observation_seen',
-      'user_correction_seen',
-      'user_interruption_seen',
-      'session_interrupted_seen',
-      'negative_feedback_seen',
-      'positive_feedback_seen',
-      'user_goal_shift_seen',
-      'hard_rule_seen',
-      'tool_failure_seen',
-      'hedging_seen',
-      'explicit_marker_seen',
-      'runtime_context_excluded',
-      'skill_context_excluded',
-      'no_priority_signal',
-    ])
-    && isEnumArray(value.cautionCodes, [
-      'no_llm_judge',
-      'rule_only',
-      'runtime_context_excluded',
-      'skill_context_excluded',
-      'no_human_user_message',
-      'limited_timeline_window',
-    ])
-    && isExperienceEvidenceRefArray(value.evidenceRefs);
-}
-
-function isExperienceProblemEvidenceRef(value: unknown): boolean {
-  if (
-    !isObjectRecord(value)
-    || typeof value.id !== 'string'
-    || typeof value.kind !== 'string'
-    || typeof value.sourceTrace !== 'string'
-    || typeof value.sessionId !== 'string'
-    || !isOptionalNonNegativeInteger(value.messageIndex)
-    || !isOptionalNonNegativeInteger(value.logicalMessageIndex)
-    || !isOptionalNonNegativeInteger(value.sourceLineIndex)
-  ) return false;
-  if (
-    value.role !== undefined
-    && !isEnumValue(value.role, ['user', 'assistant', 'tool', 'other'])
-  ) return false;
-  return [
-    value.messageUuid,
-    value.traceId,
-    value.callInstanceId,
-    value.toolUseId,
-    value.label,
-    value.snippet,
-  ].every(isOptionalString)
-    && isOptionalTimestamp(value.timestamp);
-}
-
-function isExperienceProblemPattern(value: unknown): boolean {
-  return isObjectRecord(value)
-    && typeof value.id === 'string'
-    && isEnumValue(value.bucket, [
-      'output_format',
-      'content_accuracy',
-      'missing_context',
-      'rule_violation',
-      'workflow_mismatch',
-      'tool_runtime',
-      'goal_shift',
-      'unclear',
-    ])
-    && typeof value.patternKey === 'string'
-    && isNonNegativeInteger(value.count)
-    && isNonNegativeInteger(value.sessionCount)
-    && isStringArray(value.recentSessionIds)
-    && isEnumArray(value.signalTypes, [
-      'user_correction',
-      'negative_feedback',
-      'user_interruption',
-      'hard_rule',
-      'user_goal_shift',
-      'tool_failure',
-      'workflow_mismatch',
-      'artifact_missing',
-      'observer_lifecycle_failed',
-      'orchestration_boundary_violation',
-    ])
-    && Array.isArray(value.evidenceRefs)
-    && value.evidenceRefs.every(isExperienceProblemEvidenceRef)
-    && isOptionalTimestamp(value.lastSeen);
-}
-
-function isExperienceProblemPatternArray(value: unknown): boolean {
-  return Array.isArray(value) && value.every(isExperienceProblemPattern);
-}
-
-function isExperienceAttribution(value: unknown): boolean {
-  return isObjectRecord(value)
-    && typeof value.source === 'string'
-    && isRate(value.confidence)
-    && isOptionalString(value.rawSkillRef)
-    && isOptionalString(value.pluginName)
-    && isOptionalString(value.commandName);
-}
-
-function isExperienceInvocationMetrics(value: unknown, requireSourceNeutralOutcomes = false): boolean {
-  if (!isObjectRecord(value)) return false;
-  const fields = [
-    value.durationMs,
-    value.inputTokens,
-    value.outputTokens,
-    value.cacheReadTokens,
-    value.cacheCreationTokens,
-    value.numTurns,
-    value.numToolCalls,
-    value.numToolFailures,
-  ];
-  if (!fields.every(isNonNegativeInteger)) return false;
-  if (
-    requireSourceNeutralOutcomes
-    && (
-      typeof value.tokenUsageObserved !== 'boolean'
-      ||
-      !isNonNegativeInteger(value.numToolCancelled)
-      || !isNonNegativeInteger(value.numToolUnknown)
-    )
-  ) return false;
-  if (
-    value.tokenUsageObserved !== undefined
-    && typeof value.tokenUsageObserved !== 'boolean'
-  ) return false;
-  if (value.numToolCancelled !== undefined && !isNonNegativeInteger(value.numToolCancelled)) return false;
-  if (value.numToolUnknown !== undefined && !isNonNegativeInteger(value.numToolUnknown)) return false;
-  const cancelled = typeof value.numToolCancelled === 'number' ? value.numToolCancelled : 0;
-  const unknown = typeof value.numToolUnknown === 'number' ? value.numToolUnknown : 0;
-  return (value.numToolFailures as number) + cancelled + unknown <= (value.numToolCalls as number);
-}
-
-const EXPERIENCE_INDICATOR_KEYS = [
-  'userMessageCount',
-  'userFollowUpCount',
-  'userCorrectionCount',
-  'userInterruptionCount',
-  'sessionInterruptedCount',
-  'negativeFeedbackCount',
-  'positiveFeedbackCount',
-  'userGoalShiftCount',
-  'hardRuleTextHitCount',
-  'assistantDeliverySignalCount',
-  'deliverableArtifactSignalCount',
-  'routerDownstreamCompleted',
-  'routerDownstreamFailed',
-  'selfCorrectionCount',
-  'repeatedExecutionCount',
-  'toolCallCount',
-  'toolFailureCount',
-  'highObservationCount',
-  'mediumObservationCount',
-  'hedgingCount',
-  'explicitMarkerCount',
-] as const;
-
-function isExperienceIndicators(value: unknown, requireSourceNeutralOutcomes = false): boolean {
-  return isObjectRecord(value)
-    && EXPERIENCE_INDICATOR_KEYS.every((key) => isNonNegativeInteger(value[key]))
-    && (
-      !requireSourceNeutralOutcomes
-      || (
-        isNonNegativeInteger(value.toolCancelledCount)
-        && isNonNegativeInteger(value.toolUnknownCount)
-      )
-    )
-    && (value.toolCancelledCount === undefined || isNonNegativeInteger(value.toolCancelledCount))
-    && (value.toolUnknownCount === undefined || isNonNegativeInteger(value.toolUnknownCount))
-    && (value.toolFailureCount as number)
-      + (typeof value.toolCancelledCount === 'number' ? value.toolCancelledCount : 0)
-      + (typeof value.toolUnknownCount === 'number' ? value.toolUnknownCount : 0)
-      <= (value.toolCallCount as number);
-}
-
-function isCountRecord(value: unknown): boolean {
-  return isObjectRecord(value)
-    && Object.values(value).every(isNonNegativeInteger);
-}
-
-function isExperienceTimelineScope(value: unknown): boolean {
-  if (
-    !isObjectRecord(value)
-    || value.mode !== 'skill_segment_window'
-    || !isNonNegativeInteger(value.segmentEventCount)
-    || !isNonNegativeInteger(value.previewEventCount)
-    || !isNonNegativeInteger(value.fullSessionEventCount)
-    || !isExperienceTraceRecordRangeArray(value.segmentRecordRanges)
-    || !isExperienceTraceRecordRangeArray(value.previewRecordRanges)
-    || !isExperienceTraceRecordRangeArray(value.sessionRecordRanges)
-    || typeof value.truncated !== 'boolean'
-    || !isNonNegativeInteger(value.omittedBeforeCount)
-    || !isNonNegativeInteger(value.omittedAfterCount)
-  ) return false;
-  return value.previewEventCount <= value.segmentEventCount
-    && value.segmentEventCount <= value.fullSessionEventCount
-    && value.omittedBeforeCount + value.omittedAfterCount <= value.fullSessionEventCount;
-}
-
-function isLegacyExperienceTimelineScope(value: unknown): boolean {
-  if (
-    !isObjectRecord(value)
-    || value.mode !== 'skill_segment_window'
-    || !isOptionalNonNegativeInteger(value.segmentStartRecordIndex)
-    || !isOptionalNonNegativeInteger(value.segmentEndRecordIndex)
-    || !isOptionalNonNegativeInteger(value.previewStartRecordIndex)
-    || !isOptionalNonNegativeInteger(value.previewEndRecordIndex)
-    || !isNonNegativeInteger(value.sessionStartRecordIndex)
-    || !isNonNegativeInteger(value.sessionEndRecordIndex)
-    || !isNonNegativeInteger(value.previewEventCount)
-    || !isNonNegativeInteger(value.fullSessionEventCount)
-    || typeof value.truncated !== 'boolean'
-    || !isNonNegativeInteger(value.omittedBeforeCount)
-    || !isNonNegativeInteger(value.omittedAfterCount)
-  ) return false;
-  const optionalRanges: Array<[unknown, unknown]> = [
-    [value.segmentStartRecordIndex, value.segmentEndRecordIndex],
-    [value.previewStartRecordIndex, value.previewEndRecordIndex],
-  ];
-  if (optionalRanges.some(([start, end]) =>
-    (start === undefined) !== (end === undefined)
-    || (
-      typeof start === 'number'
-      && typeof end === 'number'
-      && start > end
-    )
-  )) return false;
-  return value.sessionStartRecordIndex <= value.sessionEndRecordIndex
-    && value.previewEventCount <= value.fullSessionEventCount
-    && value.omittedBeforeCount + value.omittedAfterCount <= value.fullSessionEventCount;
-}
-
-function isExperienceTraceRecordRangeArray(value: unknown): value is ExperienceTraceRecordRange[] {
-  if (!Array.isArray(value)) return false;
-  const identities = new Set<string>();
-  for (const range of value) {
-    if (
-      !isObjectRecord(range)
-      || typeof range.traceId !== 'string'
-      || typeof range.sourceTrace !== 'string'
-      || !isNonNegativeInteger(range.startRecordIndex)
-      || !isNonNegativeInteger(range.endRecordIndex)
-      || range.startRecordIndex > range.endRecordIndex
-      || !isNonNegativeInteger(range.eventCount)
-      || range.eventCount === 0
-      || identities.has(range.traceId)
-    ) return false;
-    identities.add(range.traceId);
-  }
-  return true;
-}
-
-function isExperienceSkillSummary(value: unknown, requireToolUnknown = false): boolean {
-  if (
-    !isObjectRecord(value)
-    || typeof value.skillName !== 'string'
-    || !isNonNegativeInteger(value.invocationCount)
-    || !isNonNegativeInteger(value.sessionCount)
-    || !Array.isArray(value.sourceKinds)
-    || !value.sourceKinds.every(isObservationSourceKind)
-    || !isStringArray(value.entrypoints)
-    || !isCountRecord(value.entrypointCounts)
-    || !isCountRecord(value.attributionCounts)
-    || !isStringArray(value.pluginNames)
-    || !isStringArray(value.rawSkillRefs)
-    || !isStringArray(value.commandNames)
-    || !isCountRecord(value.toolCounts)
-    || !isTimestampRange(value.firstSeen, value.lastSeen)
-    || !isOptionalTimestampCoverage(
-      value.timestampedInvocationCount,
-      value.timestampCoverage,
-      value.invocationCount,
-    )
-    || (
-      requireToolUnknown
-      && (
-        !isNonNegativeInteger(value.timestampedInvocationCount)
-        || !isRate(value.timestampCoverage)
-      )
-    )
-    || !isNonNegativeInteger(value.reviewFirstSessionCount)
-    || !isNonNegativeInteger(value.sampleReviewSessionCount)
-    || !isExperienceIndicators(value.indicators, requireToolUnknown)
-    || !isExperienceEvidenceChain(value.evidenceChain)
-    || !isExperienceRuleFindingArray(value.ruleFindings)
-    || !isExperienceAssistiveInference(value.assistiveInference)
-    || !isExperienceProblemPatternArray(value.problemPatterns)
-    || !isStringArray(value.relatedObservationIds)
-    || !isObjectRecord(value.sourceMetadataCounts)
-  ) return false;
-  const metadataCounts = value.sourceMetadataCounts;
-  return ['channels', 'senders', 'businessActions', 'providers', 'models']
-    .every((key) => isCountRecord(metadataCounts[key]));
-}
-
-function isTimelineEvent(value: unknown): value is ExperienceTimelineEvent {
-  if (
-    !isObjectRecord(value)
-    || typeof value.id !== 'string'
-    || ![
-      'user_message',
-      'synthetic_user_event',
-      'assistant_message',
-      'model_activity',
-      'agent_activity',
-      'tool_use',
-      'tool_result',
-      'skill_context',
-      'runtime_context',
-      'lifecycle',
-      'observation',
-    ].includes(String(value.kind))
-    || typeof value.sourceTrace !== 'string'
-    || typeof value.sessionId !== 'string'
-    || !isNonNegativeInteger(value.order)
-  ) return false;
-  const optionalIndexes = [
-    value.messageIndex,
-    value.logicalMessageIndex,
-    value.sourceLineIndex,
-  ];
-  if (!optionalIndexes.every((index) => index === undefined || isNonNegativeInteger(index))) return false;
-  if (
-    value.traceRole !== undefined
-    && value.traceRole !== 'standalone'
-    && value.traceRole !== 'main'
-    && value.traceRole !== 'subagent'
-  ) return false;
-  if (
-    value.role !== undefined
-    && value.role !== 'user'
-    && value.role !== 'assistant'
-    && value.role !== 'tool'
-    && value.role !== 'other'
-  ) return false;
-  if (
-    value.toolStatus !== undefined
-    && value.toolStatus !== 'success'
-    && value.toolStatus !== 'failure'
-    && value.toolStatus !== 'cancelled'
-    && value.toolStatus !== 'unknown'
-  ) return false;
-  if (value.modelActivityKind !== undefined && value.modelActivityKind !== 'reasoning') return false;
-  if (
-    value.contentVisibility !== undefined
-    && value.contentVisibility !== 'plaintext'
-    && value.contentVisibility !== 'opaque'
-  ) return false;
-  if (
-    value.contentSource !== undefined
-    && value.contentSource !== 'summary'
-    && value.contentSource !== 'content'
-    && value.contentSource !== 'text'
-  ) return false;
-  if (
-    value.runtimeKind !== undefined
-    && value.runtimeKind !== 'session_context'
-    && value.runtimeKind !== 'execution_context'
-    && value.runtimeKind !== 'settings'
-    && value.runtimeKind !== 'goal'
-    && value.runtimeKind !== 'context_compaction'
-    && value.runtimeKind !== 'usage'
-  ) return false;
-  const optionalStrings = [
-    value.traceId,
-    value.traceLabel,
-    value.turnId,
-    value.messageUuid,
-    value.sourceType,
-    value.model,
-    value.callInstanceId,
-    value.toolUseId,
-    value.label,
-    value.snippet,
-    value.toolName,
-    value.fullText,
-  ];
-  if (
-    !optionalStrings.every((field) => field === undefined || typeof field === 'string')
-    || !isOptionalTimestamp(value.timestamp)
-    || (value.isError !== undefined && typeof value.isError !== 'boolean')
-    || (value.attachments !== undefined && !isTimelineAttachmentArray(value.attachments))
-  ) return false;
-  return value.kind !== 'tool_result'
-    || value.toolStatus === undefined
-    || value.isError === undefined
-    || value.isError === (value.toolStatus === 'failure');
-}
-
-function isTimelineAttachmentArray(value: unknown): value is NonNullable<ExperienceTimelineEvent['attachments']> {
-  return Array.isArray(value) && value.every((attachment) => (
-    isObjectRecord(attachment)
-    && (attachment.attachmentKind === 'image' || attachment.attachmentKind === 'file')
-    && typeof attachment.name === 'string'
-    && attachment.name.length > 0
-  ));
-}
-
-function isTimelineEventArray(value: unknown): value is ExperienceTimelineEvent[] {
-  return Array.isArray(value) && value.every(isTimelineEvent);
-}
-
-function isExperienceTurnSummaryArray(value: unknown): value is ExperienceTurnSummary[] {
-  return Array.isArray(value) && value.every((turn) => (
-    isObjectRecord(turn)
-    && typeof turn.turnId === 'string'
-    && isOptionalString(turn.sourceTurnId)
-    && (
-      turn.boundaryBasis === 'turn_id'
-      || turn.boundaryBasis === 'turn_lifecycle'
-      || turn.boundaryBasis === 'user_message'
-    )
-    && isOptionalString(turn.traceId)
-    && typeof turn.sourceTrace === 'string'
-    && isOptionalTimestamp(turn.startTimestamp)
-    && isOptionalTimestamp(turn.endTimestamp)
-    && (
-      turn.status === 'completed'
-      || turn.status === 'failed'
-      || turn.status === 'aborted'
-      || turn.status === 'interrupted'
-      || turn.status === 'open'
-      || turn.status === 'unknown'
-    )
-    && typeof turn.title === 'string'
-    && isStringArray(turn.eventIds)
-    && isNonNegativeInteger(turn.userMessageCount)
-    && isNonNegativeInteger(turn.assistantMessageCount)
-    && isNonNegativeInteger(turn.toolCallCount)
-    && isNonNegativeInteger(turn.toolFailureCount)
-  ));
-}
-
-function isTimelineTree(value: unknown): value is ExperienceTimelineTree {
-  if (
-    !isObjectRecord(value)
-    || typeof value.sessionId !== 'string'
-    || !isTimelineEventArray(value.main)
-    || !Array.isArray(value.branches)
-  ) return false;
-  return value.branches.every((branch) => {
-    if (
-      !isObjectRecord(branch)
-      || typeof branch.id !== 'string'
-      || typeof branch.label !== 'string'
-      || typeof branch.sessionId !== 'string'
-      || !isOptionalString(branch.traceId)
-      || typeof branch.sourceTrace !== 'string'
-      || (
-        branch.traceRole !== 'main'
-        && branch.traceRole !== 'subagent'
-        && branch.traceRole !== 'standalone'
-      )
-      || !isTimelineEventArray(branch.events)
-    ) return false;
-    if (branch.attachTo === undefined) return true;
-    if (
-      !isObjectRecord(branch.attachTo)
-      || !isOptionalString(branch.attachTo.traceId)
-      || typeof branch.attachTo.sourceTrace !== 'string'
-    ) return false;
-    return (branch.attachTo.messageIndex === undefined || isNonNegativeInteger(branch.attachTo.messageIndex))
-      && (branch.attachTo.callInstanceId === undefined || typeof branch.attachTo.callInstanceId === 'string')
-      && (branch.attachTo.toolUseId === undefined || typeof branch.attachTo.toolUseId === 'string')
-      && (branch.attachTo.label === undefined || typeof branch.attachTo.label === 'string');
-  });
-}
-
-function isExperienceChecklistItem(value: unknown): boolean {
-  return isObjectRecord(value)
-    && typeof value.key === 'string'
-    && typeof value.label === 'string'
-    && isEnumValue(value.status, [
-      'passed',
-      'failed',
-      'unknown',
-      'not_declared',
-      'not_applicable',
-      'degraded',
-    ])
-    && isEnumValue(value.contribution, [
-      'blocking',
-      'attention',
-      'informational',
-      'positive',
-      'neutral',
-    ])
-    && typeof value.reason === 'string'
-    && isExperienceEvidenceRefArray(value.evidenceRefs)
-    && isEnumValue(value.source, ['deterministic_rule', 'llm_soft', 'manual'])
-    && isOptionalString(value.suggestionKey);
-}
-
-function isExperienceSessionStoryGoalSlice(value: unknown): boolean {
-  return isObjectRecord(value)
-    && typeof value.id === 'string'
-    && isNonNegativeInteger(value.order)
-    && isStringArray(value.skillNames)
-    && isTimestampRange(value.startTimestamp, value.endTimestamp)
-    && isEnumValue(value.reasonCode, [
-      'skill_segment_boundary',
-      'explicit_user_goal_shift',
-      'default_session_slice',
-    ])
-    && isOptionalString(value.inferredUserGoal)
-    && isExperienceEvidenceRefArray(value.evidenceRefs);
-}
-
-function isExperienceSessionStorySubagentDispatch(value: unknown): boolean {
-  if (
-    !isObjectRecord(value)
-    || typeof value.id !== 'string'
-    || !isNonNegativeInteger(value.order)
-    || typeof value.branchId !== 'string'
-    || typeof value.childSessionId !== 'string'
-    || typeof value.traceId !== 'string'
-    || typeof value.label !== 'string'
-    || typeof value.sourceTrace !== 'string'
-    || !isNonNegativeInteger(value.eventCount)
-    || !isExperienceEvidenceRefArray(value.evidenceRefs)
-  ) return false;
-  if (value.attachTo === undefined) return true;
-  return isObjectRecord(value.attachTo)
-    && isOptionalNonNegativeInteger(value.attachTo.messageIndex)
-    && isOptionalString(value.attachTo.callInstanceId)
-    && isOptionalString(value.attachTo.toolUseId)
-    && isOptionalString(value.attachTo.label);
-}
-
-function isExperienceSessionStorySkillLink(value: unknown): boolean {
-  return isObjectRecord(value)
-    && typeof value.id === 'string'
-    && isNonNegativeInteger(value.order)
-    && typeof value.skillName === 'string'
-    && isEnumValue(value.role, ['router', 'executor', 'mixed', 'unknown'])
-    && isStringArray(value.invocationIds)
-    && isStringArray(value.goalSliceIds)
-    && isExperienceEvidenceRefArray(value.evidenceRefs);
-}
-
-function isExperienceSessionStoryGraphNode(value: unknown): boolean {
-  return isObjectRecord(value)
-    && typeof value.id === 'string'
-    && typeof value.label === 'string'
-    && isEnumValue(value.kind, [
-      'user_goal',
-      'skill_invocation',
-      'subagent_branch',
-      'tool_execution',
-      'delivery',
-      'user_feedback',
-      'goal_shift',
-    ])
-    && isEnumValue(value.status, ['ok', 'attention', 'unknown', 'degraded', 'not_applicable'])
-    && (value.role === undefined || isEnumValue(value.role, ['router', 'executor', 'mixed', 'unknown']))
-    && isOptionalString(value.detailNodeId);
-}
-
-function isExperienceSessionStoryGraphEdge(value: unknown): boolean {
-  return isObjectRecord(value)
-    && typeof value.fromId === 'string'
-    && typeof value.toId === 'string'
-    && typeof value.label === 'string';
-}
-
-function isExperienceSessionStoryNode(value: unknown): boolean {
-  return isObjectRecord(value)
-    && typeof value.id === 'string'
-    && isNonNegativeInteger(value.order)
-    && isEnumValue(value.kind, [
-      'user_goal',
-      'skill_invocation',
-      'subagent_branch',
-      'tool_execution',
-      'delivery',
-      'user_feedback',
-      'goal_shift',
-    ])
-    && typeof value.label === 'string'
-    && isEnumValue(value.status, ['ok', 'attention', 'unknown', 'degraded', 'not_applicable'])
-    && typeof value.text === 'string'
-    && isExperienceEvidenceRefArray(value.evidenceRefs);
-}
-
-function isExperienceSessionStoryAnswer(value: unknown): boolean {
-  return isObjectRecord(value)
-    && isEnumValue(value.key, ['goal_satisfaction', 'declared_behavior_fit', 'user_feeling'])
-    && typeof value.label === 'string'
-    && isEnumValue(value.status, ['ok', 'attention', 'unknown', 'degraded', 'not_applicable'])
-    && isEnumValue(value.reason, [
-      'data_degraded',
-      'blocking_failed',
-      'attention_accumulated',
-      'unknown_dominant',
-      'all_passed',
-      'not_applicable',
-    ])
-    && isStringArray(value.sourceItemKeys)
-    && typeof value.text === 'string'
-    && isExperienceEvidenceRefArray(value.evidenceRefs)
-    && Array.isArray(value.checklistItems)
-    && value.checklistItems.every(isExperienceChecklistItem);
-}
-
-function isExperienceGoalEvidenceRef(value: unknown): boolean {
-  return isObjectRecord(value)
-    && isEnumValue(value.kind, ['user_message', 'goal_slice', 'llm_goal'])
-    && isOptionalString(value.goalSliceId)
-    && (value.evidenceRef === undefined || isExperienceEvidenceRef(value.evidenceRef))
-    && isOptionalString(value.label);
-}
-
-function isExperienceSkillSegment(value: unknown): boolean {
-  if (
-    !isObjectRecord(value)
-    || typeof value.id !== 'string'
-    || !isNonNegativeInteger(value.order)
-    || typeof value.skillName !== 'string'
-    || !isEnumValue(value.skillType, [
-      'router',
-      'delegation',
-      'executor',
-      'advisory',
-      'workflow_owner',
-      'unknown',
-    ])
-    || (
-      value.skillTypeSource !== undefined
-      && !isEnumValue(value.skillTypeSource, ['frontmatter', 'trace', 'unknown'])
-    )
-    || (
-      value.declaredSkillType !== undefined
-      && !isEnumValue(value.declaredSkillType, [
-        'router',
-        'delegation',
-        'executor',
-        'advisory',
-        'workflow_owner',
-        'unknown',
-      ])
-    )
-    || (
-      value.traceInferredSkillType !== undefined
-      && !isEnumValue(value.traceInferredSkillType, [
-        'router',
-        'delegation',
-        'executor',
-        'advisory',
-        'workflow_owner',
-        'unknown',
-      ])
-    )
-    || !isEnumValue(value.episodeRole, [
-      'main_executor',
-      'router',
-      'delegator',
-      'supporting',
-      'observer',
-    ])
-    || !isStringArray(value.skillInvocationIds)
-    || !isOptionalNonNegativeInteger(value.startMessageIndex)
-    || !isOptionalNonNegativeInteger(value.endMessageIndex)
-    || !isTimestampRange(value.startTimestamp, value.endTimestamp)
-    || !Array.isArray(value.typeSpecificChecklist)
-    || !value.typeSpecificChecklist.every(isExperienceChecklistItem)
-    || !isExperienceEvidenceRefArray(value.evidenceRefs)
-  ) return false;
-  if (
-    value.messageRanges !== undefined
-    && (
-      !Array.isArray(value.messageRanges)
-      || value.messageRanges.some((range) =>
-        !isObjectRecord(range)
-        || !isNonNegativeInteger(range.startMessageIndex)
-        || !isNonNegativeInteger(range.endMessageIndex)
-        || range.startMessageIndex > range.endMessageIndex
-        || !isOptionalString(range.traceId)
-        || !isOptionalString(range.sourceTrace)
-        || !isOptionalString(range.sessionId)
-      )
-    )
-  ) return false;
-  if (
-    typeof value.startMessageIndex === 'number'
-    && typeof value.endMessageIndex === 'number'
-    && value.startMessageIndex > value.endMessageIndex
-  ) return false;
-  if (value.runtimeAssessment === undefined) return true;
-  return isObjectRecord(value.runtimeAssessment)
-    && isOptionalString(value.runtimeAssessment.goalSatisfaction)
-    && isOptionalString(value.runtimeAssessment.declaredBehaviorFit)
-    && isOptionalString(value.runtimeAssessment.artifactGoalMatch)
-    && isOptionalString(value.runtimeAssessment.userFeeling);
-}
-
-function isExperienceOrchestrationEdge(value: unknown): boolean {
-  return isObjectRecord(value)
-    && typeof value.id === 'string'
-    && typeof value.episodeId === 'string'
-    && isEnumValue(value.edgeKind, ['internal_skill', 'external_child_session'])
-    && isOptionalString(value.parentSkillSegmentId)
-    && isOptionalString(value.executorSkillSegmentId)
-    && isOptionalString(value.childSessionId)
-    && (value.runnerStartedRef === undefined || isExperienceEvidenceRef(value.runnerStartedRef))
-    && (value.runnerCompletedRef === undefined || isExperienceEvidenceRef(value.runnerCompletedRef))
-    && (value.notificationRef === undefined || isExperienceEvidenceRef(value.notificationRef))
-    && isEnumValue(value.status, ['started', 'completed', 'failed', 'unknown'])
-    && isExperienceEvidenceRefArray(value.evidenceRefs);
-}
-
-function isExperienceFeedbackAttribution(value: unknown): boolean {
-  return isObjectRecord(value)
-    && isOptionalString(value.skillName)
-    && isOptionalString(value.skillSegmentId)
-    && isEnumValue(value.attributionRole, ['primary_fault', 'downstream_related', 'context_only'])
-    && isEnumValue(value.reasonCode, [
-      'object_match',
-      'promise_match',
-      'action_match',
-      'orchestration_edge',
-      'episode_context',
-    ])
-    && isExperienceEvidenceRefArray(value.evidenceRefs);
-}
-
-function isExperienceFeedbackSignal(value: unknown): boolean {
-  return isObjectRecord(value)
-    && typeof value.id === 'string'
-    && isNonNegativeInteger(value.order)
-    && isEnumValue(value.type, [
-      'correction',
-      'follow_up',
-      'frustration',
-      'interruption',
-      'positive',
-      'unknown',
-    ])
-    && typeof value.text === 'string'
-    && isOptionalString(value.targetObject)
-    && isEnumValue(value.sourceWindow, [
-      'session',
-      'episode',
-      'skill_invocation',
-      'downstream_child',
-    ])
-    && isExperienceEvidenceRef(value.evidenceRef)
-    && (
-      value.canonicalAttributions === undefined
-      || (
-        Array.isArray(value.canonicalAttributions)
-        && value.canonicalAttributions.every(isExperienceFeedbackAttribution)
-      )
-    )
-    && Array.isArray(value.attributions)
-    && value.attributions.every(isExperienceFeedbackAttribution);
-}
-
-function isExperienceEpisodeArtifact(value: unknown): boolean {
-  return isObjectRecord(value)
-    && isEnumValue(value.kind, [
-      'path',
-      'url',
-      'document',
-      'code',
-      'execution_window',
-      'unknown',
-    ])
-    && typeof value.label === 'string'
-    && isOptionalString(value.pathOrUrl)
-    && isEnumValue(value.artifactGoalMatch, ['passed', 'failed', 'unknown'])
-    && isExperienceEvidenceRef(value.evidenceRef);
-}
-
-function isExperienceEpisodeOutcome(value: unknown): boolean {
-  return isObjectRecord(value)
-    && isEnumValue(value.closure, ['closed', 'unresolved', 'abandoned', 'unknown'])
-    && Array.isArray(value.artifacts)
-    && value.artifacts.every(isExperienceEpisodeArtifact)
-    && isExperienceReviewPriority(value.verdict)
-    && isOptionalString(value.acceptanceCriteria);
-}
-
-function isExperienceEpisode(value: unknown): boolean {
-  return isObjectRecord(value)
-    && typeof value.id === 'string'
-    && isNonNegativeInteger(value.order)
-    && typeof value.sessionId === 'string'
-    && isOptionalString(value.primaryGoal)
-    && Array.isArray(value.goalEvidenceRefs)
-    && value.goalEvidenceRefs.every(isExperienceGoalEvidenceRef)
-    && isTimestampRange(value.startTimestamp, value.endTimestamp)
-    && (value.startRef === undefined || isExperienceEvidenceRef(value.startRef))
-    && (value.endRef === undefined || isExperienceEvidenceRef(value.endRef))
-    && isEnumValue(value.boundaryReason, [
-      'goal_shift',
-      'checkpoint_or_subagent',
-      'downstream_closed',
-      'session_end',
-    ])
-    && Array.isArray(value.skillSegments)
-    && value.skillSegments.every(isExperienceSkillSegment)
-    && Array.isArray(value.orchestrationEdges)
-    && value.orchestrationEdges.every(isExperienceOrchestrationEdge)
-    && Array.isArray(value.feedbackSignals)
-    && value.feedbackSignals.every(isExperienceFeedbackSignal)
-    && isExperienceEpisodeOutcome(value.outcome);
-}
-
-function isExperienceStoryContext(value: unknown): value is ExperienceStoryContext {
-  return isObjectRecord(value)
-    && typeof value.id === 'string'
-    && typeof value.sessionGroupKey === 'string'
-    && Array.isArray(value.goalSlices)
-    && value.goalSlices.every(isExperienceSessionStoryGoalSlice)
-    && Array.isArray(value.subagentDispatches)
-    && value.subagentDispatches.every(isExperienceSessionStorySubagentDispatch)
-    && Array.isArray(value.episodes)
-    && value.episodes.every(isExperienceEpisode);
-}
-
-function isExperienceSessionStory(value: unknown, compact: boolean): boolean {
-  if (
-    !isObjectRecord(value)
-    || value.schemaVersion !== 1
-    || (compact ? typeof value.contextRef !== 'string' : !isOptionalString(value.contextRef))
-    || typeof value.summary !== 'string'
-    || !isNonNegativeInteger(value.invocationCount)
-    || !isNonNegativeInteger(value.goalSliceCount)
-    || !isNonNegativeInteger(value.branchCount)
-    || !isNonNegativeInteger(value.progressUpdateCount)
-    || !isNonNegativeInteger(value.finalDeliverySignalCount)
-    || !isStringArray(value.mainlineNodeIds)
-    || !Array.isArray(value.skillLinks)
-    || !value.skillLinks.every(isExperienceSessionStorySkillLink)
-    || !isObjectRecord(value.graph)
-    || !Array.isArray(value.graph.nodes)
-    || !value.graph.nodes.every(isExperienceSessionStoryGraphNode)
-    || !Array.isArray(value.graph.edges)
-    || !value.graph.edges.every(isExperienceSessionStoryGraphEdge)
-    || !Array.isArray(value.nodes)
-    || !value.nodes.every(isExperienceSessionStoryNode)
-    || !Array.isArray(value.answers)
-    || !value.answers.every(isExperienceSessionStoryAnswer)
-  ) return false;
-  if (compact) {
-    return value.goalSlices === undefined
-      && value.subagentDispatches === undefined
-      && value.episodes === undefined;
-  }
-  return Array.isArray(value.goalSlices)
-    && value.goalSlices.every(isExperienceSessionStoryGoalSlice)
-    && Array.isArray(value.subagentDispatches)
-    && value.subagentDispatches.every(isExperienceSessionStorySubagentDispatch)
-    && (
-      value.episodes === undefined
-      || (Array.isArray(value.episodes) && value.episodes.every(isExperienceEpisode))
-    );
-}
-
-function isExperienceReviewerReportStep(value: unknown): boolean {
-  return isObjectRecord(value)
-    && isNonNegativeInteger(value.order)
-    && typeof value.label === 'string'
-    && isEnumValue(value.status, ['ok', 'attention', 'unknown', 'degraded', 'not_applicable'])
-    && typeof value.text === 'string'
-    && isExperienceEvidenceRefArray(value.evidenceRefs);
-}
-
-function isExperienceReviewerReportFinding(value: unknown): boolean {
-  if (
-    !isObjectRecord(value)
-    || typeof value.id !== 'string'
-    || typeof value.judgmentId !== 'string'
-    || !isEnumValue(value.source, ['deterministic_rule', 'llm_soft', 'manual'])
-    || !isEnumValue(value.level, ['attention', 'possible_false_positive', 'note'])
-    || typeof value.title !== 'string'
-    || typeof value.body !== 'string'
-    || typeof value.ruleSource !== 'string'
-    || typeof value.ruleVersion !== 'string'
-    || !isExperienceEvidenceRefArray(value.evidenceRefs)
-    || !isObjectRecord(value.reviewStateRef)
-    || value.reviewStateRef.targetType !== 'reviewer_judgment'
-    || typeof value.reviewStateRef.targetId !== 'string'
-  ) return false;
-  return [
-    value.reviewStateRef.verdict,
-    value.reviewStateRef.reason,
-    value.reviewStateRef.note,
-  ].every(isOptionalString)
-    && isOptionalTimestamp(value.reviewStateRef.reviewedAt);
-}
-
-function isExperienceReviewerMetrics(value: unknown, requireSourceNeutralOutcomes: boolean): boolean {
-  if (!isObjectRecord(value)) return false;
-  const countKeys = [
-    'toolCallCount',
-    'toolFailureCount',
-    'userMessageCount',
-    'userFollowUpCount',
-    'assistantDeliverySignalCount',
-    'deliverableArtifactSignalCount',
-    'routerDownstreamCompleted',
-    'routerDownstreamFailed',
-    'assistantProgressUpdateCount',
-    'selfCorrectionCount',
-    'repeatedExecutionCount',
-    'finalDeliverySignalCount',
-    'traceEventCount',
-  ];
-  if (
-    !countKeys.every((key) => isNonNegativeInteger(value[key]))
-    || (
-      requireSourceNeutralOutcomes
-      && (
-        !isNonNegativeInteger(value.toolCancelledCount)
-        || !isNonNegativeInteger(value.toolUnknownCount)
-      )
-    )
-    || (
-      value.toolCancelledCount !== undefined
-      && !isNonNegativeInteger(value.toolCancelledCount)
-    )
-    || (
-      value.toolUnknownCount !== undefined
-      && !isNonNegativeInteger(value.toolUnknownCount)
-    )
-    || !isObjectRecord(value.tokenUsage)
-  ) return false;
-  const tokenKeys = [
-    'inputTokens',
-    'outputTokens',
-    'cacheReadTokens',
-    'cacheCreationTokens',
-  ];
-  const tokenUsage = value.tokenUsage as Record<string, unknown>;
-  const observedInvocationCount = tokenUsage.observedInvocationCount;
-  const invocationCount = tokenUsage.invocationCount;
-  const coverage = tokenUsage.coverage;
-  const hasCoverage = observedInvocationCount !== undefined
-    || invocationCount !== undefined
-    || coverage !== undefined;
-  return tokenKeys.every((key) => isNonNegativeInteger(tokenUsage[key]))
-    && tokenUsage.attribution === 'skill_segment'
-    && (
-      !hasCoverage
-      || (
-        isNonNegativeInteger(observedInvocationCount)
-        && isNonNegativeInteger(invocationCount)
-        && isRate(coverage)
-        && (observedInvocationCount as number) <= (invocationCount as number)
-        && Math.abs(
-          (coverage as number)
-          - (
-            (invocationCount as number) > 0
-              ? (observedInvocationCount as number) / (invocationCount as number)
-              : 1
-          ),
-        ) <= 0.0001
-      )
-    )
-    && (!requireSourceNeutralOutcomes || hasCoverage)
-    && (value.toolFailureCount as number)
-      + (typeof value.toolCancelledCount === 'number' ? value.toolCancelledCount : 0)
-      + (typeof value.toolUnknownCount === 'number' ? value.toolUnknownCount : 0)
-      <= (value.toolCallCount as number);
-}
-
-function isExperienceReviewerReport(value: unknown, compact: boolean): boolean {
-  if (
-    !isObjectRecord(value)
-    || value.schemaVersion !== 1
-    || !isEnumValue(value.mode, ['deterministic_milestone_1', 'deterministic_session_story'])
-    || !isTimestamp(value.generatedAt)
-    || typeof value.title !== 'string'
-    || typeof value.summary !== 'string'
-    || !isObjectRecord(value.scope)
-    || !isEnumValue(value.scope.kind, ['single_skill_single_goal', 'degraded_complex'])
-    || !isStringArray(value.scope.reasonCodes)
-    || !Array.isArray(value.chainSteps)
-    || !value.chainSteps.every(isExperienceReviewerReportStep)
-    || !Array.isArray(value.findings)
-    || !value.findings.every(isExperienceReviewerReportFinding)
-    || !isExperienceReviewerMetrics(value.oneLookMetrics, compact)
-    || !isStringArray(value.authorSuggestions)
-    || !isExperienceEvidenceRefArray(value.traceLinks)
-  ) return false;
-  if (compact) {
-    return value.sessionStory === undefined && value.sessionStoryRef === 'session';
-  }
-  return isExperienceSessionStory(value.sessionStory, false)
-    && (value.sessionStoryRef === undefined || value.sessionStoryRef === 'session');
-}
-
-/**
- * Converts a hydrated report into the wire format used by inbox JSON.
- * Timeline evidence is stored once per logical trace; consumers hydrate the
- * invocation/session views through normalizeObservationExperienceReport().
- */
-export function compactObservationExperienceReport(
-  report: ObservationExperienceReport,
-): PersistedObservationExperienceReport {
-  const hydrated = hydrateExperienceTimelines(report);
-  return {
-    ...hydrated,
-    invocations: hydrated.invocations.map((invocation) => {
-      const { timeline, ...persisted } = invocation;
-      return {
-        ...persisted,
-        timelineRef: invocation.timelineRef
-          ?? timelineRefForSessionGroup(invocation.sessionGroupKey),
-        timelineEventIds: invocation.timelineEventIds
-          ?? timeline.map((event) => event.id),
-      };
-    }),
-    sessions: hydrated.sessions.map((session) => {
-      const { timelinePreview, sessionStory, reviewerReport } = session;
-      const persisted = omitProperties(session, [
-        'attributedEventIds',
-        'timelinePreview',
-        'fullSessionTimeline',
-        'timelineTree',
-        'sessionStory',
-        'reviewerReport',
-      ]);
-      const contextRef = sessionStory?.contextRef;
-      return {
-        ...persisted,
-        timelineRef: session.timelineRef
-          ?? timelineRefForSessionGroup(session.sessionId),
-        timelinePreviewEventIds: session.timelinePreviewEventIds
-          ?? timelinePreview.map((event) => event.id),
-        sessionStory: sessionStory && contextRef
-          ? compactSessionStory(sessionStory, contextRef)
-          : undefined,
-        reviewerReport: reviewerReport && sessionStory && contextRef
-          ? {
-              ...omitSessionStory(reviewerReport),
-              sessionStoryRef: 'session',
-            }
-          : undefined,
-      };
-    }),
-  };
-}
-
-function hydrateExperienceTimelines(
-  report: ObservationExperienceReport,
-): ObservationExperienceReport {
-  const timelineById = new Map(report.traceTimelines.map((timeline) => [timeline.id, timeline]));
-  const storyContextById = new Map(report.storyContexts.map((context) => [context.id, context]));
-  const flattenedById = new Map<string, ExperienceTimelineEvent[]>();
-  const eventByTimelineId = new Map<string, Map<string, ExperienceTimelineEvent>>();
-  const flattenedTimeline = (ref: string): ExperienceTimelineEvent[] => {
-    const cached = flattenedById.get(ref);
-    if (cached) return cached;
-    const timeline = timelineById.get(ref);
-    if (!timeline) return [];
-    const flattened = flattenTimelineTree(timeline.tree);
-    flattenedById.set(ref, flattened);
-    eventByTimelineId.set(ref, new Map(flattened.map((event) => [event.id, event])));
-    return flattened;
-  };
-  const selectEvents = (
-    ref: string,
-    ids: string[] | undefined,
-    fallback: ExperienceTimelineEvent[],
-  ): ExperienceTimelineEvent[] => {
-    if (!timelineById.has(ref)) return fallback;
-    if (!ids) return fallback.length > 0 ? fallback : flattenedTimeline(ref);
-    const byId = eventByTimelineId.get(ref)
-      ?? new Map(flattenedTimeline(ref).map((event) => [event.id, event]));
-    return ids.flatMap((id) => {
-      const event = byId.get(id);
-      return event ? [event] : [];
-    });
-  };
-
-  const invocations = report.invocations.map((invocation): ExperienceInvocation => {
-    const timelineRef = invocation.timelineRef ?? timelineRefForSessionGroup(invocation.sessionGroupKey);
-    const timelineEventIds = invocation.timelineEventIds
-      ?? invocation.timeline.map((event) => event.id);
-    return {
-      ...invocation,
-      timelineRef,
-      timelineEventIds,
-      timeline: selectEvents(timelineRef, timelineEventIds, invocation.timeline),
-    };
-  });
-  const invocationById = new Map(invocations.map((invocation) => [invocation.id, invocation]));
-  const sessions = report.sessions.map((session): ExperienceSessionSummary => {
-    const firstInvocation = session.invocationIds
-      .map((id) => invocationById.get(id))
-      .find((invocation): invocation is ExperienceInvocation => Boolean(invocation));
-    const timelineRef = session.timelineRef
-      ?? firstInvocation?.timelineRef
-      ?? timelineRefForSessionGroup(firstInvocation?.sessionGroupKey ?? session.sessionId);
-    const storedTimeline = timelineById.get(timelineRef);
-    const fullSessionTimeline = storedTimeline
-      ? flattenedTimeline(timelineRef)
-      : session.fullSessionTimeline;
-    const turns = session.turns.length > 0
-      ? session.turns
-      : reconstructExperienceTurns(fullSessionTimeline);
-    const timelinePreviewEventIds = session.timelinePreviewEventIds
-      ?? session.timelinePreview.map((event) => event.id);
-    const timelinePreview = timelinePreviewEventIds.length > 0
-      ? selectEvents(timelineRef, timelinePreviewEventIds, session.timelinePreview)
-      : fullSessionTimeline.slice(0, TIMELINE_PREVIEW_EVENT_LIMIT);
-    const attributedEventIds = unique(session.invocationIds.flatMap((id) => {
-      const invocation = invocationById.get(id);
-      return invocation?.timelineEventIds ?? invocation?.timeline.map((event) => event.id) ?? [];
-    }));
-    const contextRef = session.sessionStory?.contextRef
-      ?? storyContextRefForSessionGroup(firstInvocation?.sessionGroupKey ?? session.sessionId);
-    const storyContext = storyContextById.get(contextRef);
-    const sessionStory = session.sessionStory
-      ? {
-          ...session.sessionStory,
-          contextRef,
-          goalSlices: storyContext?.goalSlices ?? session.sessionStory.goalSlices,
-          subagentDispatches: storyContext?.subagentDispatches
-            ?? session.sessionStory.subagentDispatches,
-          episodes: storyContext?.episodes ?? session.sessionStory.episodes ?? [],
-        }
-      : undefined;
-    return {
-      ...session,
-      timelineRef,
-      attributedEventIds,
-      timelinePreviewEventIds: timelinePreview.map((event) => event.id),
-      timelinePreview,
-      fullSessionTimeline,
-      turns,
-      timelineTree: storedTimeline?.tree ?? session.timelineTree,
-      sessionStory,
-      reviewerReport: session.reviewerReport && sessionStory
-        ? {
-            ...session.reviewerReport,
-            sessionStory: session.reviewerReport.sessionStoryRef === 'session'
-              ? sessionStory
-              : session.reviewerReport.sessionStory,
-          }
-        : session.reviewerReport,
-    };
-  });
-
-  return {
-    ...report,
-    schemaVersion: OBSERVATION_EXPERIENCE_SCHEMA_VERSION,
-    invocations,
-    sessions,
-  };
-}
-
-function validateExperienceReferences(
-  report: ObservationExperienceReport,
-  strict: boolean,
-): boolean {
-  if (
-    !isObjectRecord(report.meta)
-    || report.meta.sessionCount !== report.sessions.length
-    || report.meta.skillCount !== report.skills.length
-    || report.meta.invocationCount !== report.invocations.length
-    || report.meta.goalSliceCount !== report.goalSlices.length
-  ) return false;
-
-  const timelineEventsByRef = new Map<string, Set<string>>();
-  const timelineByRef = new Map<string, ExperienceTraceTimeline>();
-  const timelineSessionGroupKeys = new Set<string>();
-  for (const timeline of report.traceTimelines) {
-    if (
-      timelineEventsByRef.has(timeline.id)
-      || timelineSessionGroupKeys.has(timeline.sessionGroupKey)
-    ) return false;
-    if (
-      timeline.tree.sessionId !== timeline.sessionId
-      || (strict && !traceTimelineStructureIsConsistent(timeline))
-    ) return false;
-    const rawEvents = [
-      ...timeline.tree.main,
-      ...timeline.tree.branches.flatMap((branch) => branch.events),
-    ];
-    if (
-      strict
-      && (
-        rawEvents.some((event) => typeof event.traceId !== 'string')
-        || rawEvents.some((event) =>
-          (event.kind === 'tool_use' || event.kind === 'tool_result')
-          && typeof event.callInstanceId !== 'string'
-        )
-        || timeline.tree.branches.some((branch) =>
-          typeof branch.traceId !== 'string'
-          || (branch.attachTo !== undefined && typeof branch.attachTo.traceId !== 'string')
-        )
-      )
-    ) return false;
-    if (new Set(rawEvents.map((event) => event.id)).size !== rawEvents.length) return false;
-    const events = flattenTimelineTree(timeline.tree);
-    if (timeline.eventCount !== events.length) return false;
-    timelineEventsByRef.set(timeline.id, new Set(events.map((event) => event.id)));
-    timelineByRef.set(timeline.id, timeline);
-    timelineSessionGroupKeys.add(timeline.sessionGroupKey);
-  }
-
-  const goalSliceIds = new Set<string>();
-  const goalSliceById = new Map<string, ExperienceGoalSlice>();
-  for (const goalSlice of report.goalSlices) {
-    if (
-      !isExperienceGoalSlice(goalSlice)
-      || (
-        strict
-        && (
-          typeof goalSlice.traceId !== 'string'
-          || typeof goalSlice.timestampObserved !== 'boolean'
-        )
-      )
-      || goalSliceIds.has(goalSlice.id)
-    ) return false;
-    goalSliceIds.add(goalSlice.id);
-    goalSliceById.set(goalSlice.id, goalSlice as unknown as ExperienceGoalSlice);
-  }
-
-  const invocationById = new Map<string, ExperienceInvocation>();
-  const referencedGoalSliceIds = new Set<string>();
-  for (const invocation of report.invocations) {
-    if (invocationById.has(invocation.id)) return false;
-    invocationById.set(invocation.id, invocation);
-    if (
-      strict
-      && (
-        countRecordTotal(invocation.toolCounts) !== invocation.metrics.numToolCalls
-        || !toolOutcomeCountsMatch(
-          invocation.metrics,
-          invocation.indicators,
-        )
-      )
-    ) return false;
-    if (!goalSliceIds.has(invocation.goalSliceId)) return false;
-    const goalSlice = goalSliceById.get(invocation.goalSliceId);
-    if (
-      goalSlice
-      && (
-        goalSlice.skillName !== invocation.skillName
-        || goalSlice.sessionId !== invocation.sessionId
-        || goalSlice.sourceTrace !== invocation.sourceTrace
-        || goalSlice.startTimestamp !== invocation.startTimestamp
-        || goalSlice.endTimestamp !== invocation.endTimestamp
-        || (
-          strict
-          && (
-            goalSlice.traceId !== invocation.traceId
-            || goalSlice.timestampObserved !== invocation.timestampObserved
-          )
-        )
-      )
-    ) return false;
-    referencedGoalSliceIds.add(invocation.goalSliceId);
-    if (strict && (!invocation.timelineRef || !invocation.timelineEventIds)) return false;
-    if (!invocation.timelineRef) {
-      if (invocation.timeline.length === 0 && (invocation.timelineEventIds?.length ?? 0) > 0) return false;
-      continue;
-    }
-    const eventIds = timelineEventsByRef.get(invocation.timelineRef);
-    if (!eventIds) return false;
-    if (timelineByRef.get(invocation.timelineRef)?.sessionGroupKey !== invocation.sessionGroupKey) return false;
-    if (
-      invocation.timelineEventIds
-      && new Set(invocation.timelineEventIds).size !== invocation.timelineEventIds.length
-    ) return false;
-    if (invocation.timelineEventIds?.some((id) => !eventIds.has(id))) return false;
-  }
-  if (referencedGoalSliceIds.size !== goalSliceIds.size) return false;
-
-  const sessionIds = new Set<string>();
-  const invocationReferenceCounts = new Map(
-    report.invocations.map((invocation) => [invocation.id, 0]),
-  );
-  const referencedTimelineIds = new Set<string>();
-  const referencedStoryContextIds = new Set<string>();
-  for (const session of report.sessions) {
-    if (sessionIds.has(session.id)) return false;
-    sessionIds.add(session.id);
-    if (
-      session.invocationIds.length === 0
-      || new Set(session.invocationIds).size !== session.invocationIds.length
-      || session.invocationIds.some((id) => !invocationById.has(id))
-    ) return false;
-    if (session.goalSliceIds.some((id) => !goalSliceIds.has(id))) return false;
-    const sessionInvocations = session.invocationIds
-      .map((id) => invocationById.get(id))
-      .filter((value): value is ExperienceInvocation => Boolean(value));
-    const sessionGroupKeys = new Set(sessionInvocations.map((invocation) => invocation.sessionGroupKey));
-    const invocationGoalSliceIds = new Set(sessionInvocations.map((invocation) => invocation.goalSliceId));
-    if (
-      sessionInvocations.some((invocation) => invocation.skillName !== session.skillName)
-      || sessionGroupKeys.size !== 1
-      || session.goalSliceIds.length !== invocationGoalSliceIds.size
-      || session.goalSliceIds.some((id) => !invocationGoalSliceIds.has(id))
-    ) return false;
-    for (const invocation of sessionInvocations) {
-      invocationReferenceCounts.set(
-        invocation.id,
-        (invocationReferenceCounts.get(invocation.id) ?? 0) + 1,
-      );
-    }
-    if (strict) {
-      const timestampedInvocations = sessionInvocations.filter(invocationTimestampObserved);
-      const expectedStartTimestamp = minString(
-        timestampedInvocations.map((invocation) => invocation.startTimestamp),
-      ) ?? sessionInvocations[0]?.startTimestamp;
-      const expectedEndTimestamp = maxString(
-        timestampedInvocations.map((invocation) => invocation.endTimestamp),
-      ) ?? sessionInvocations[0]?.endTimestamp;
-      if (
-        sessionInvocations.some((invocation) => invocation.timelineRef !== session.timelineRef)
-      ) return false;
-      const invocationIndicators = sumIndicators(
-        sessionInvocations.map((invocation) => invocation.indicators),
-      );
-      const expectedIndicators = enrichRouterDownstreamIndicators({
-        ...session,
-        indicators: invocationIndicators,
-      });
-      const expectedReviewPriorityScore = scoreForIndicators(expectedIndicators);
-      const expectedReviewPriority = session.reviewerReport
-        ? priorityForReviewerFindings({
-            ...session,
-            indicators: expectedIndicators,
-            reviewPriorityScore: expectedReviewPriorityScore,
-          }, session.reviewerReport.findings)
-        : priorityForScore(expectedReviewPriorityScore);
-      if (
-        !indicatorRecordsEqual(session.indicators, expectedIndicators)
-        || session.reviewPriorityScore !== expectedReviewPriorityScore
-        || session.reviewPriority !== expectedReviewPriority
-        || !sameStringArray(
-          session.reviewBasisCodes,
-          basisCodesForIndicators(expectedIndicators),
-        )
-        || session.startTimestamp !== expectedStartTimestamp
-        || session.endTimestamp !== expectedEndTimestamp
-        || session.timestampedInvocationCount !== timestampedInvocations.length
-        || session.timestampCoverage !== timestampedInvocations.length / sessionInvocations.length
-      ) return false;
-      const sessionGroupKey = sessionInvocations[0]?.sessionGroupKey;
-      if (
-        sessionGroupKey
-        && timelineByRef.get(session.timelineRef ?? '')?.sessionGroupKey !== sessionGroupKey
-      ) return false;
-    }
-    if (strict && (!session.timelineRef || !session.timelinePreviewEventIds)) return false;
-    if (session.timelineRef) {
-      referencedTimelineIds.add(session.timelineRef);
-      const eventIds = timelineEventsByRef.get(session.timelineRef);
-      const timeline = timelineByRef.get(session.timelineRef);
-      if (!eventIds) return false;
-      if (
-        session.timelinePreviewEventIds
-        && new Set(session.timelinePreviewEventIds).size !== session.timelinePreviewEventIds.length
-      ) return false;
-      if (session.timelinePreviewEventIds?.some((id) => !eventIds.has(id))) return false;
-      if (
-        strict
-        && (
-          timeline?.sessionId !== session.sessionId
-          || !timelineScopeMatches(
-            session,
-            sessionInvocations,
-            timeline,
-          )
-        )
-      ) return false;
-    } else if (
-      session.fullSessionTimeline.length === 0
-      && session.timelinePreview.length === 0
-      && (session.timelinePreviewEventIds?.length ?? 0) > 0
-    ) {
-      return false;
-    }
-    if (session.sessionStory?.contextRef) {
-      referencedStoryContextIds.add(session.sessionStory.contextRef);
-    }
-    if (
-      strict
-      && session.reviewerReport
-      && !reviewerMetricsMatch(
-        session.reviewerReport,
-        session,
-        sessionInvocations,
-        timelineByRef.get(session.timelineRef ?? ''),
-      )
-    ) return false;
-  }
-  if (
-    Array.from(invocationReferenceCounts.values()).some((count) => count !== 1)
-    || (strict && referencedTimelineIds.size !== timelineEventsByRef.size)
-  ) return false;
-
-  const storyContextIds = new Set<string>();
-  const storyContextById = new Map<string, ExperienceStoryContext>();
-  const storyContextSessionGroupKeys = new Set<string>();
-  for (const context of report.storyContexts) {
-    if (
-      storyContextIds.has(context.id)
-      || storyContextSessionGroupKeys.has(context.sessionGroupKey)
-    ) return false;
-    storyContextIds.add(context.id);
-    storyContextById.set(context.id, context);
-    storyContextSessionGroupKeys.add(context.sessionGroupKey);
-  }
-  const skillNames = new Set<string>();
-  for (const skill of report.skills) {
-    if (
-      !isExperienceSkillSummary(skill, strict)
-      || skillNames.has(skill.skillName)
-    ) return false;
-    if (strict) {
-      const skillInvocations = report.invocations.filter(
-        (invocation) => invocation.skillName === skill.skillName,
-      );
-      const skillSessions = report.sessions.filter(
-        (session) => session.skillName === skill.skillName,
-      );
-      const timestampedInvocations = skillInvocations.filter(invocationTimestampObserved);
-      const expectedFirstSeen = minString(
-        timestampedInvocations.map((invocation) => invocation.startTimestamp),
-      ) ?? skillSessions[0]?.startTimestamp;
-      const expectedLastSeen = maxString(
-        timestampedInvocations.map((invocation) => invocation.endTimestamp),
-      ) ?? skillSessions[0]?.endTimestamp;
-      if (
-        skill.invocationCount !== skillInvocations.length
-        || skill.sessionCount !== skillSessions.length
-        || !indicatorRecordsEqual(
-          skill.indicators,
-          sumIndicators(skillSessions.map((session) => session.indicators)),
-        )
-        || !countRecordsEqual(
-          skill.toolCounts,
-          sumRecordCounts(skillInvocations.map((invocation) => invocation.toolCounts)),
-        )
-        || skill.firstSeen !== expectedFirstSeen
-        || skill.lastSeen !== expectedLastSeen
-        || skill.timestampedInvocationCount !== timestampedInvocations.length
-        || skill.timestampCoverage !== (
-          skillInvocations.length > 0
-            ? timestampedInvocations.length / skillInvocations.length
-            : 0
-        )
-        || skill.reviewFirstSessionCount !== skillSessions.filter(
-          (session) => session.reviewPriority === 'review_first',
-        ).length
-        || skill.sampleReviewSessionCount !== skillSessions.filter(
-          (session) => session.reviewPriority === 'sample_review',
-        ).length
-      ) return false;
-    }
-    skillNames.add(skill.skillName);
-  }
-  if (
-    report.sessions.some((session) => !skillNames.has(session.skillName))
-    || report.invocations.some((invocation) => !skillNames.has(invocation.skillName))
-    || (strict && referencedStoryContextIds.size !== storyContextIds.size)
-  ) return false;
-  return report.sessions.every((session) => {
-    const contextRef = session.sessionStory?.contextRef;
-    if (strict && session.sessionStory && !contextRef) return false;
-    if (contextRef && !storyContextIds.has(contextRef)) return false;
-    if (strict && contextRef) {
-      const firstInvocation = session.invocationIds
-        .map((id) => invocationById.get(id))
-        .find((value): value is ExperienceInvocation => Boolean(value));
-      const storyInvocations = firstInvocation
-        ? report.invocations.filter(
-            (invocation) => invocation.sessionGroupKey === firstInvocation.sessionGroupKey,
-          )
-        : [];
-      const context = storyContextById.get(contextRef);
-      const timeline = timelineByRef.get(session.timelineRef ?? '');
-      if (
-        firstInvocation
-        && context?.sessionGroupKey !== firstInvocation.sessionGroupKey
-      ) return false;
-      if (
-        session.sessionStory
-        && context
-        && timeline
-        && !sessionStoryStructureIsConsistent(
-          session.sessionStory,
-          context,
-          storyInvocations,
-          timeline,
-        )
-      ) return false;
-    }
-    if (
-      strict
-      && session.reviewerReport
-      && session.reviewerReport.sessionStoryRef !== 'session'
-      && !isObjectRecord((session.reviewerReport as unknown as Record<string, unknown>).sessionStory)
-    ) return false;
-    return session.reviewerReport?.sessionStoryRef !== 'session' || Boolean(session.sessionStory);
-  });
-}
-
-function sessionStoryStructureIsConsistent(
-  story: ExperienceSessionStory,
-  context: ExperienceStoryContext,
-  invocations: ExperienceInvocation[],
-  timeline: ExperienceTraceTimeline,
-): boolean {
-  const timelineEvents = flattenTimelineTree(timeline.tree);
-  const expectedGoalSliceIds = new Set(invocations.map((invocation) => invocation.goalSliceId));
-  const contextGoalSliceIds = context.goalSlices.map((goalSlice) => goalSlice.id);
-  if (
-    story.invocationCount !== invocations.length
-    || story.goalSliceCount !== expectedGoalSliceIds.size
-    || contextGoalSliceIds.length !== expectedGoalSliceIds.size
-    || new Set(contextGoalSliceIds).size !== contextGoalSliceIds.length
-    || contextGoalSliceIds.some((id) => !expectedGoalSliceIds.has(id))
-    || story.branchCount !== timeline.tree.branches.length
-    || story.progressUpdateCount !== timelineEvents.filter(isAssistantProgressUpdateEvent).length
-    || story.finalDeliverySignalCount !== timelineEvents.filter(isAssistantDeliveryEvent).length
-  ) return false;
-  if (!storyContextEpisodesAreConsistent(context, invocations, timeline.sessionId)) return false;
-
-  const branchById = new Map(timeline.tree.branches.map((branch) => [branch.id, branch]));
-  if (
-    context.subagentDispatches.length !== branchById.size
-    || new Set(context.subagentDispatches.map((dispatch) => dispatch.id)).size
-      !== context.subagentDispatches.length
-    || new Set(context.subagentDispatches.map((dispatch) => dispatch.branchId)).size
-      !== context.subagentDispatches.length
-    || context.subagentDispatches.some((dispatch) => {
-      const branch = branchById.get(dispatch.branchId);
-      return !branch
-        || dispatch.childSessionId !== branch.sessionId
-        || dispatch.traceId !== (branch.traceId ?? branch.sourceTrace)
-        || dispatch.sourceTrace !== branch.sourceTrace
-        || dispatch.eventCount !== branch.events.length;
-    })
-  ) return false;
-
-  const invocationById = new Map(invocations.map((invocation) => [invocation.id, invocation]));
-  const invocationReferenceCounts = new Map(
-    invocations.map((invocation) => [invocation.id, 0]),
-  );
-  const skillLinkIds = new Set<string>();
-  for (const link of story.skillLinks) {
-    if (
-      skillLinkIds.has(link.id)
-      || new Set(link.invocationIds).size !== link.invocationIds.length
-      || link.invocationIds.some((id) => !invocationById.has(id))
-    ) return false;
-    skillLinkIds.add(link.id);
-    const linkedInvocations = link.invocationIds
-      .map((id) => invocationById.get(id))
-      .filter((value): value is ExperienceInvocation => Boolean(value));
-    const linkedGoalSliceIds = new Set(
-      linkedInvocations.map((invocation) => invocation.goalSliceId),
-    );
-    if (
-      linkedInvocations.some((invocation) => invocation.skillName !== link.skillName)
-      || link.goalSliceIds.length !== linkedGoalSliceIds.size
-      || link.goalSliceIds.some((id) => !linkedGoalSliceIds.has(id))
-    ) return false;
-    for (const invocation of linkedInvocations) {
-      invocationReferenceCounts.set(
-        invocation.id,
-        (invocationReferenceCounts.get(invocation.id) ?? 0) + 1,
-      );
-    }
-  }
-  if ([...invocationReferenceCounts.values()].some((count) => count !== 1)) return false;
-
-  const nodeIds = story.nodes.map((node) => node.id);
-  const nodeIdSet = new Set(nodeIds);
-  const graphNodeIds = story.graph.nodes.map((node) => node.id);
-  const expectedGraphNodeIds = new Set([...nodeIds, ...skillLinkIds]);
-  if (
-    nodeIdSet.size !== nodeIds.length
-    || new Set(story.mainlineNodeIds).size !== story.mainlineNodeIds.length
-    || story.mainlineNodeIds.some((id) => !nodeIdSet.has(id))
-    || new Set(graphNodeIds).size !== graphNodeIds.length
-    || graphNodeIds.length !== expectedGraphNodeIds.size
-    || graphNodeIds.some((id) => !expectedGraphNodeIds.has(id))
-    || story.graph.nodes.some((node) =>
-      node.detailNodeId !== undefined && !nodeIdSet.has(node.detailNodeId)
-    )
-    || story.graph.edges.some((edge) =>
-      !expectedGraphNodeIds.has(edge.fromId) || !expectedGraphNodeIds.has(edge.toId)
-    )
-  ) return false;
-
-  return true;
-}
-
-function timelineScopeMatches(
-  session: ExperienceSessionSummary,
-  invocations: ExperienceInvocation[],
-  timeline: ExperienceTraceTimeline | undefined,
-): boolean {
-  if (!timeline) return false;
-  const events = flattenTimelineTree(timeline.tree);
-  const eventById = new Map(events.map((event) => [event.id, event]));
-  const invocationEventIds = new Set(
-    invocations.flatMap((invocation) => invocation.timelineEventIds ?? []),
-  );
-  const previewEventIds = session.timelinePreviewEventIds ?? [];
-  if (previewEventIds.some((id) => !invocationEventIds.has(id))) return false;
-  const invocationEvents = [...invocationEventIds]
-    .map((id) => eventById.get(id))
-    .filter((event): event is ExperienceTimelineEvent => Boolean(event));
-  if (invocationEvents.length !== invocationEventIds.size) return false;
-  const previewEvents = previewEventIds
-    .map((id) => eventById.get(id))
-    .filter((event): event is ExperienceTimelineEvent => Boolean(event));
-  const eventPositionById = new Map(events.map((event, index) => [event.id, index]));
-  const previewPositions = previewEventIds
-    .map((id) => eventPositionById.get(id))
-    .filter((index): index is number => typeof index === 'number');
-  const omittedBeforeCount = previewPositions.length > 0 ? Math.min(...previewPositions) : 0;
-  const omittedAfterCount = previewPositions.length > 0
-    ? events.length - 1 - Math.max(...previewPositions)
-    : 0;
-  const scope = session.timelineScope;
-  return sameStringSet(session.attributedEventIds, invocationEventIds)
-    && scope.segmentEventCount === invocationEventIds.size
-    && scope.previewEventCount === previewEventIds.length
-    && scope.fullSessionEventCount === events.length
-    && traceRecordRangesEqual(scope.segmentRecordRanges, traceRecordRanges(invocationEvents))
-    && traceRecordRangesEqual(scope.previewRecordRanges, traceRecordRanges(previewEvents))
-    && traceRecordRangesEqual(scope.sessionRecordRanges, traceRecordRanges(events))
-    && scope.omittedBeforeCount === omittedBeforeCount
-    && scope.omittedAfterCount === omittedAfterCount
-    && scope.truncated === (
-      invocationEventIds.size > previewEventIds.length
-      || omittedBeforeCount > 0
-      || omittedAfterCount > 0
-    );
-}
-
-function traceRecordRangesEqual(
-  left: ExperienceTraceRecordRange[],
-  right: ExperienceTraceRecordRange[],
-): boolean {
-  return left.length === right.length
-    && left.every((range, index) =>
-      range.traceId === right[index].traceId
-      && range.sourceTrace === right[index].sourceTrace
-      && range.startRecordIndex === right[index].startRecordIndex
-      && range.endRecordIndex === right[index].endRecordIndex
-      && range.eventCount === right[index].eventCount
-    );
-}
-
-function storyContextEpisodesAreConsistent(
-  context: ExperienceStoryContext,
-  invocations: ExperienceInvocation[],
-  sessionId: string,
-): boolean {
-  const goalSliceIds = new Set(context.goalSlices.map((goalSlice) => goalSlice.id));
-  const invocationById = new Map(invocations.map((invocation) => [invocation.id, invocation]));
-  const episodeIds = new Set<string>();
-  for (const episode of context.episodes) {
-    if (episodeIds.has(episode.id) || episode.sessionId !== sessionId) return false;
-    episodeIds.add(episode.id);
-
-    const skillSegmentById = new Map<string, ExperienceSkillSegment>();
-    for (const segment of episode.skillSegments) {
-      if (
-        skillSegmentById.has(segment.id)
-        || new Set(segment.skillInvocationIds).size !== segment.skillInvocationIds.length
-        || segment.skillInvocationIds.some((id) => invocationById.get(id)?.skillName !== segment.skillName)
-        || Date.parse(segment.startTimestamp) < Date.parse(episode.startTimestamp)
-        || Date.parse(segment.endTimestamp) > Date.parse(episode.endTimestamp)
-      ) return false;
-      skillSegmentById.set(segment.id, segment);
-    }
-
-    const edgeIds = new Set<string>();
-    for (const edge of episode.orchestrationEdges) {
-      const parent = edge.parentSkillSegmentId
-        ? skillSegmentById.get(edge.parentSkillSegmentId)
-        : undefined;
-      const executor = edge.executorSkillSegmentId
-        ? skillSegmentById.get(edge.executorSkillSegmentId)
-        : undefined;
-      if (
-        edgeIds.has(edge.id)
-        || edge.episodeId !== episode.id
-        || (edge.parentSkillSegmentId !== undefined && !parent)
-        || (edge.executorSkillSegmentId !== undefined && !executor)
-        || (
-          edge.edgeKind === 'internal_skill'
-          && (!parent || !executor || parent.id === executor.id)
-        )
-      ) return false;
-      edgeIds.add(edge.id);
-    }
-
-    const signalIds = new Set<string>();
-    for (const signal of episode.feedbackSignals) {
-      if (signalIds.has(signal.id)) return false;
-      signalIds.add(signal.id);
-      const attributionGroups = [
-        signal.attributions,
-        signal.canonicalAttributions ?? [],
-      ];
-      for (const attribution of attributionGroups.flat()) {
-        if (!attribution.skillSegmentId) continue;
-        const segment = skillSegmentById.get(attribution.skillSegmentId);
-        if (!segment || (attribution.skillName && attribution.skillName !== segment.skillName)) {
-          return false;
-        }
-      }
-    }
-
-    if (episode.goalEvidenceRefs.some((ref) =>
-      ref.kind === 'goal_slice'
-        ? !ref.goalSliceId || !goalSliceIds.has(ref.goalSliceId)
-        : ref.goalSliceId !== undefined && !goalSliceIds.has(ref.goalSliceId)
-    )) return false;
-  }
-  return true;
-}
-
-function traceTimelineStructureIsConsistent(timeline: ExperienceTraceTimeline): boolean {
-  const branches = timeline.tree.branches;
-  if (new Set(branches.map((branch) => branch.id)).size !== branches.length) return false;
-  if (timeline.tree.main.some((event) => event.sessionId !== timeline.sessionId)) return false;
-  const mainTraceIds = new Set(
-    timeline.tree.main
-      .map((event) => event.traceId)
-      .filter((value): value is string => Boolean(value)),
-  );
-  const mainSourceTraces = new Set(timeline.tree.main.map((event) => event.sourceTrace));
-  if (mainTraceIds.size > 1 || mainSourceTraces.size > 1) return false;
-
-  for (const branch of branches) {
-    if (branch.events.some((event) =>
-      event.sessionId !== timeline.sessionId
-      || (branch.traceId !== undefined && event.traceId !== branch.traceId)
-      || event.sourceTrace !== branch.sourceTrace
-      || event.traceRole !== branch.traceRole
-    )) return false;
-    if (
-      branch.attachTo
-      && !timeline.tree.main.some((event) =>
-        (
-          branch.attachTo?.traceId !== undefined
-            ? event.traceId === branch.attachTo.traceId
-            : event.sourceTrace === branch.attachTo?.sourceTrace
-        )
-        && event.sourceTrace === branch.attachTo?.sourceTrace
-        && (
-          branch.attachTo.messageIndex === undefined
-          || event.messageIndex === branch.attachTo.messageIndex
-        )
-        && (
-          branch.attachTo.callInstanceId === undefined
-          || event.callInstanceId === branch.attachTo.callInstanceId
-        )
-        && (
-          branch.attachTo.toolUseId === undefined
-          || event.toolUseId === branch.attachTo.toolUseId
-        )
-        && (
-          branch.attachTo.label === undefined
-          || event.label === branch.attachTo.label
-          || event.toolName === branch.attachTo.label
-        )
-      )
-    ) return false;
-  }
-  return true;
-}
-
-type ToolOutcomeSummary = Pick<
-  ExperienceReviewIndicators,
-  'toolCallCount' | 'toolFailureCount' | 'toolCancelledCount' | 'toolUnknownCount'
->;
-
-type InvocationToolOutcomeSummary = Pick<
-  ExperienceInvocation['metrics'],
-  'numToolCalls' | 'numToolFailures' | 'numToolCancelled' | 'numToolUnknown'
->;
-
-function toolOutcomeCountsMatch(
-  left: ToolOutcomeSummary | InvocationToolOutcomeSummary,
-  right: ToolOutcomeSummary,
-): boolean {
-  const normalized = 'numToolCalls' in left
-    ? {
-        toolCallCount: left.numToolCalls,
-        toolFailureCount: left.numToolFailures,
-        toolCancelledCount: left.numToolCancelled,
-        toolUnknownCount: left.numToolUnknown,
-      }
-    : left;
-  return normalized.toolCallCount === right.toolCallCount
-    && normalized.toolFailureCount === right.toolFailureCount
-    && (normalized.toolCancelledCount ?? 0) === (right.toolCancelledCount ?? 0)
-    && (normalized.toolUnknownCount ?? 0) === (right.toolUnknownCount ?? 0);
-}
-
-function indicatorRecordsEqual(
-  left: ExperienceReviewIndicators,
-  right: ExperienceReviewIndicators,
-): boolean {
-  const keys = [
-    ...EXPERIENCE_INDICATOR_KEYS,
-    'toolCancelledCount',
-    'toolUnknownCount',
-  ] as const;
-  return keys.every((key) => (left[key] ?? 0) === (right[key] ?? 0));
-}
-
-function reviewerMetricsMatch(
-  reviewer: ExperienceReviewerReport,
-  session: ExperienceSessionSummary,
-  invocations: ExperienceInvocation[],
-  timeline: ExperienceTraceTimeline | undefined,
-): boolean {
-  const metrics = reviewer.oneLookMetrics;
-  const expectedUsage = sumTokenUsage(invocations);
-  const timelineEvents = timeline ? flattenTimelineTree(timeline.tree) : [];
-  const hydratedSession = {
-    ...session,
-    fullSessionTimeline: timelineEvents,
-    timelineTree: timeline?.tree,
-  };
-  return toolOutcomeCountsMatch(
-    metrics,
-    session.indicators,
-  )
-    && metrics.userMessageCount === session.indicators.userMessageCount
-    && metrics.assistantDeliverySignalCount === session.indicators.assistantDeliverySignalCount
-    && metrics.deliverableArtifactSignalCount === session.indicators.deliverableArtifactSignalCount
-    && metrics.routerDownstreamCompleted === session.indicators.routerDownstreamCompleted
-    && metrics.routerDownstreamFailed === session.indicators.routerDownstreamFailed
-    && metrics.assistantProgressUpdateCount === assistantProgressUpdateEvents(hydratedSession).length
-    && metrics.selfCorrectionCount === session.indicators.selfCorrectionCount
-    && metrics.repeatedExecutionCount === session.indicators.repeatedExecutionCount
-    && metrics.finalDeliverySignalCount === assistantFinalDeliveryEvents(hydratedSession).length
-    && metrics.traceEventCount === timelineEvents.length
-    && metrics.tokenUsage.inputTokens === expectedUsage.inputTokens
-    && metrics.tokenUsage.outputTokens === expectedUsage.outputTokens
-    && metrics.tokenUsage.cacheReadTokens === expectedUsage.cacheReadTokens
-    && metrics.tokenUsage.cacheCreationTokens === expectedUsage.cacheCreationTokens
-    && metrics.tokenUsage.observedInvocationCount === expectedUsage.observedInvocationCount
-    && metrics.tokenUsage.invocationCount === expectedUsage.invocationCount
-    && metrics.tokenUsage.coverage === expectedUsage.coverage;
-}
-
-function countRecordTotal(value: Record<string, number>): number {
-  return sumSafeCounts(...Object.values(value));
-}
-
-function countRecordsEqual(left: Record<string, number>, right: Record<string, number>): boolean {
-  const keys = new Set([...Object.keys(left), ...Object.keys(right)]);
-  return [...keys].every((key) => (left[key] ?? 0) === (right[key] ?? 0));
-}
-
-function sameStringArray(left: string[], right: string[]): boolean {
-  return left.length === right.length && left.every((value, index) => value === right[index]);
-}
-
-function sameStringSet(left: string[], right: Set<string>): boolean {
-  return left.length === right.size
-    && new Set(left).size === left.length
-    && left.every((value) => right.has(value));
-}
-
-function compactSessionStory(
-  story: ExperienceSessionStory,
-  contextRef: string,
-): PersistedExperienceSessionStory {
-  const persisted = omitProperties(story, [
-    'goalSlices',
-    'subagentDispatches',
-    'episodes',
-  ]);
-  return { ...persisted, contextRef };
-}
-
-function omitSessionStory(
-  report: ExperienceReviewerReport,
-): Omit<ExperienceReviewerReport, 'sessionStory' | 'sessionStoryRef'> {
-  return omitProperties(report, ['sessionStory', 'sessionStoryRef']);
-}
-
-function omitProperties<T extends object, K extends keyof T>(
-  value: T,
-  keys: readonly K[],
-): Omit<T, K> {
-  const copy = { ...value };
-  for (const key of keys) delete copy[key];
-  return copy;
 }
 
 function relatedObservationItems(segment: SkillSegment, items: ObservationInboxItem[]): ObservationInboxItem[] {
